FLCatch constructor

Make an FLCatch object.

Catch data for a single species or stock unit is handled by the FLCatch class. Data is separated as landings and discards by age, in numbers, with the corresponding mean weights at age.

FLCatch(object, ...)

catch.sel(object, ...) <- value

discards.ratio(object, ...)

lrevenue(object, ...)

# S4 method for FLCatch
landings.n(object)

# S4 method for FLCatch,numeric
landings.n(object) <- value

# S4 method for FLCatch,numeric
landings.wt(object) <- value

# S4 method for FLCatch,numeric
discards.n(object) <- value

# S4 method for FLCatch,numeric
discards.wt(object) <- value

# S4 method for FLCatch,numeric
catch.sel(object) <- value

# S4 method for FLCatch,numeric
price(object) <- value

# S4 method for FLCatch,numeric
catch.q(object) <- value

# S4 method for FLCatch
landings(object)

# S4 method for FLCatch
discards(object)

# S4 method for FLCatch
catch(object)

# S4 method for FLCatch
catch.n(object)

# S4 method for FLCatch
catch.wt(object)

# S4 method for FLQuant
FLCatch(object, ...)

# S4 method for missing
FLCatch(object, ...)

# S4 method for FLCatch
lrevenue(object)

# S4 method for FLCatch
landings.sel(object)

# S4 method for FLCatch
discards.sel(object)

# S4 method for FLCatch
discards.ratio(object)

# S4 method for FLCatch,missing
plot(x, y, ...)

Arguments

object

Either an FLQuant (to determine the size of the FLQuant slots) or missing

...

Other things

value

Replacement value

x

FLCatch

y

missing

Value

An FLCatch object

Details

Make an FLCatch object.

This is class is used inside FLFishery to store the catches of a single stock or species caught by that fleet.

Slots

catch.q

Parameters of the catchability function, (FLPar).

catch.sel

Selectivity at age as proportions over fully-selected ages, (FLQuant).

desc

Description of the data contents and origin, (character).

discards.n

Discards at age in numbers, (FLQuant).

discards.wt

Mean weight-at-age in the discards, (FLQuant).

name

Name of the object, e.g. species or stock code, (character).

landings.n

Landings at age in numbers, (FLQuant).

landings.wt

Mean weight-at-age in the landings, (FLQuant).

price

Mean price by age per unit of weight, (FLQuant).

range

Ranges of age and years, plusgroup, (numeric).

Validity

Length of dimensions 2:5

All FLQuant slots must share dimensions 2 to 5.

iter dim of length 1 or N

The 6th dimension in all FLQuant and FLPar slots must be 1 or N, where N is the same value for the whole object.

Length of dimensions 2:5

All FLQuant slots must share dimensions 2 to 5.

You can inspect the class validity function by using getValidity(getClassDef('FLCatch'))

Accessors

All slots in the class have accessor and replacement methods defined that allow retrieving and substituting individual slots.

The values passed for replacement need to be of the class of that slot. A numeric vector can also be used when replacing FLQuant slots, and the vector will be used to substitute the values in the slot, but not its other attributes.

Constructor

A construction method exists for this class that can take named arguments for any of its slots. All slots are then created to match the requirements of the class validity. If an unnamed FLQuant object is provided, this is used for sizing but not stored in any slot.

Methods

Methods exist for various calculations based on values stored in the class:

landings

Total landings as sum on 'age' of landings.n times landings.wt.

discards

Total discards as sum on 'age' of discards.n times discards.wt.

landings.sel

Selectivity at age in the landings as proportions over fully-selected ages, (FLQuant).

discards.sel

Selectivity at age in the discards as proportions over fully-selected ages, (FLQuant).

catch.n

Catch at age in numbers as landings.n plus discards.n.

catch.wt

Weighted average of landings.wt and discards.wt.

catch

Total catch as sum of landings and discards.

discards.ratio

Proportion at age of discards in catch.

plot

Standard plot for the FLCatch class.

See also

Author

The FLR Team

Iago Mosqueira, EC JRC.

Examples

data(ple4)

# EXTRACT data from FLCore ple4, fake prices
fca <- FLCatch(name='PLE', desc='All NS PLE catches',
  landings.n=landings.n(ple4), landings.wt=landings.wt(ple4),
  discards.n=discards.n(ple4), discards.wt=discards.wt(ple4),
  price=landings.wt(ple4) * 23, catch.q=FLPar(q=1),
  catch.sel=catch.sel(ple4))

# Calculations
landings(fca)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> age   1957   1958   1959   1960   1961   1962   1963   1964   1965   1966  
#>   all  70926  74157  78178  88764  85267  90305 103162 111121 105424  98334
#>      year
#> age   1967   1968   1969   1970   1971   1972   1973   1974   1975   1976  
#>   all 103947 121020 122661 111783 117301 130443 133768 115181  94458 122166
#>      year
#> age   1977   1978   1979   1980   1981   1982   1983   1984   1985   1986  
#>   all 108524 128366 119827 150640 151304 145669 143690 162681 182374 166633
#>      year
#> age   1987   1988   1989   1990   1991   1992   1993   1994   1995   1996  
#>   all 155005 168118 187666 174414 147844 134793 141800 126195 109681  93642
#>      year
#> age   1997   1998   1999   2000   2001   2002   2003   2004   2005   2006  
#>   all  82875  74217  97340 100318  66596  87138  74770  82529  61501  62161
#>      year
#> age   2007   2008   2009   2010   2011   2012   2013   2014   2015   2016  
#>   all  59713  64574  68842  74246  74437  82307  97123  86424  91033  86381
#>      year
#> age   2017  
#>   all  84750
#> 
#> units:  t 

catch.n(fca)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  1957       1958       1959       1960       1961       1962      
#>   1    42702.66   72733.05   99992.13   97351.36  105215.69   65815.54
#>   2    40141.28   71294.94  120597.60  152894.50  135019.66  147178.83
#>   3    79353.02   52031.60   80618.90  116614.20  137289.10  129967.50
#>   4    56560.05   67368.68   37763.40   51480.30   72242.52   89048.40
#>   5    31888.26   35848.98   43350.57   24004.78   32458.56   46321.31
#>   6    10988.02   21828.30   23811.03   27315.98   14742.89   20597.37
#>   7    12049.59    8568.27   14322.21   13940.69   15632.32    8738.10
#>   8     9595.80    8610.92    5999.88    9719.19    9045.32    9787.41
#>   9     5495.25    7050.64    6932.95    4592.23    6532.14    5770.30
#>   10   11779.09   14773.43   16725.71   15503.41   12505.12   13593.85
#>     year
#> age  1963       1964       1965       1966       1967       1968      
#>   1    51938.98  167733.27   41740.48   39663.31   37965.74   52595.95
#>   2   111598.54  116858.90  446985.00  111166.40   98310.59   82427.96
#>   3   161466.70  126250.50  123478.00  462675.40  120682.10  105022.10
#>   4    83925.51   90727.70   61372.16   61390.10  259964.00   71629.90
#>   5    57929.89   52829.48   53769.44   35187.61   35582.32  157422.80
#>   6    30973.68   38431.54   32417.30   30721.82   20186.83   21932.60
#>   7    12304.87   17096.54   18841.81   15318.96   15625.57   11270.73
#>   8     5372.20    7257.27    9369.66   10117.96    8730.53    9465.05
#>   9     6709.53    3823.86    4806.34    5832.41    6397.31    5731.52
#>   10   16653.29   19738.48   16512.27   13434.04   12784.66   14154.41
#>     year
#> age  1969       1970       1971       1972       1973       1974      
#>   1   109457.72  117963.12   73200.74   61075.31  245992.44  216251.72
#>   2    87678.20  133995.30  126945.00   83724.50   76932.20  308599.00
#>   3    74641.70   66343.10  100832.40  108372.60   77394.22   66810.75
#>   4    59702.90   41494.61   39796.45   67677.23   74297.75   47873.80
#>   5    45774.70   39552.22   27756.81   26517.95   42478.64   40611.35
#>   6   103358.62   30184.47   25591.45   17801.04   16061.13   22360.23
#>   7    12474.47   57327.07   17180.59   15407.22   10556.57    8600.26
#>   8     6747.59    7215.49   33998.65   10747.18    9710.27    6360.36
#>   9     6186.04    4365.02    4767.89   23096.60    7321.75    6455.12
#>   10   15527.56   17101.29   16488.17   15528.78   26611.03   22141.89
#>     year
#> age  1975       1976       1977       1978       1979       1980      
#>   1   187654.33  186977.31  311012.79  260241.69  237416.19  228048.89
#>   2   237186.50  168020.90  142113.10  222931.10  198319.50  218847.00
#>   3   232622.90  162809.20  113342.50   96339.80  152671.10  139741.72
#>   4    36792.38  128298.60   96958.20   69719.70   56458.10   80323.48
#>   5    23001.21   17915.05   70724.47   58168.61   40407.81   28827.79
#>   6    18900.43   11094.15    9943.65   41726.33   31902.36   19874.09
#>   7    11126.04    9921.34    6354.82    5572.73   21132.34   15387.55
#>   8     4998.26    6528.95    5846.88    3587.01    2988.53   11113.54
#>   9     4078.97    3088.62    3861.27    3341.05    2028.66    1676.81
#>   10   17837.63   13249.71    9737.37    8318.56    7618.36    6699.37
#>     year
#> age  1981       1982       1983       1984       1985       1986      
#>   1   185495.15  378858.07  312379.15  330326.34  468120.57 1083225.66
#>   2   269603.30  249567.40  462798.80  315003.20  294917.80  417134.40
#>   3   158997.60  195147.39  177915.40  330543.20  227034.40  213479.40
#>   4    65082.75   69630.68   91099.41   93427.50  186569.17  128385.20
#>   5    35922.17   27199.44   29609.48   41553.29   46043.95   96350.40
#>   6    13526.22   16783.02   12738.97   14250.85   20729.69   23032.71
#>   7     9994.94    7064.21    8864.51    6893.69    7680.85   10371.03
#>   8     8074.76    5264.18    3930.15    5312.93    4138.25    4255.10
#>   9     6106.59    4501.20    3172.50    2486.08    3255.40    2437.65
#>   10    5886.32    8191.05    8233.28    6983.05    5857.16    6299.68
#>     year
#> age  1987       1988       1989       1990       1991       1992      
#>   1   442682.83  374165.61  245185.87  205383.51  187958.37  168039.58
#>   2  1047256.40  478975.90  438711.90  298563.10  258742.30  242447.10
#>   3   302191.00  746685.00  331436.00  307676.00  223193.20  201697.80
#>   4   114725.07  148044.90  337985.90  152615.20  153467.20  115796.60
#>   5    65923.14   56419.88   71205.46  168825.14   80728.20   82733.51
#>   6    46662.01   31129.37   26393.55   33025.90   78235.21   38323.92
#>   7    10445.49   19869.41   12392.43    9827.96   12398.30   31918.82
#>   8     5248.06    4881.37    8448.05    5182.11    4630.49    6547.11
#>   9     2503.27    2916.79    2412.42    4215.04    3060.81    3024.36
#>   10    6660.50    6327.28    5119.13    3977.21    5138.26    5860.65
#>     year
#> age  1993       1994       1995       1996       1997       1998      
#>   1   104636.60   93290.14  126854.53  105190.15  272197.35   91657.58
#>   2   211487.60  128362.70  128617.10  232415.20  255916.60  701954.20
#>   3   183104.50  156029.90  106282.90  129090.50  242378.70  221681.40
#>   4   100871.90   88971.80   79720.43   56087.87   58913.90   85780.00
#>   5    60965.29   51632.74   44764.95   38876.43   25344.58   24336.34
#>   6    40691.96   30198.53   25052.05   21493.70   18509.83   11435.31
#>   7    16523.26   17190.20   12590.23   10906.74    9435.01    7208.62
#>   8    15502.77    6601.74    6586.87    5404.86    4914.63    3754.64
#>   9     3569.78    6301.80    2542.41    2953.18    2644.70    2236.10
#>   10    5512.56    4370.61    4990.73    4174.69    4409.92    4065.62
#>     year
#> age  1999       2000       2001       2002       2003       2004      
#>   1    91337.92  128525.79   97495.00  273411.66   91796.79  235514.85
#>   2   186960.90  138479.50  187737.70  172176.50  531815.50  146919.67
#>   3   505281.00  147232.10  142364.40  199413.20  136174.20  311226.00
#>   4    76360.70  242206.10   92615.80   77923.40   74693.40   43158.15
#>   5    34618.15   32822.43  109946.10   41589.14   34002.05   32763.40
#>   6     9804.85   12638.84   12252.56   47176.75   20056.13   16254.67
#>   7     3720.12    3078.29    4697.26    5248.95   19344.91    7219.69
#>   8     2527.38    1433.52    1455.38    2162.48    1819.76    5516.38
#>   9     1623.54    1209.85     716.29     593.44     673.88     541.68
#>   10    3199.13    2395.19    1899.89    1503.05    1324.69    1330.89
#>     year
#> age  2005       2006       2007       2008       2009       2010      
#>   1   172458.80  148426.56  190996.81  155534.36  167161.84  208142.93
#>   2   271017.30  175205.50  177494.30  259601.30  177602.40  141436.20
#>   3    84388.60  183320.40  127413.40  120064.40  165611.40  125432.90
#>   4   117804.80   37951.03   79448.20   52002.37   56248.90  101973.80
#>   5    19322.44   52637.29   16996.50   36841.46   25518.30   29618.53
#>   6    14652.44    8757.52   26391.73    9157.63   19803.51   13736.46
#>   7     5814.47    6184.97    4385.02   14378.47    5198.22   11568.37
#>   8     2303.57    2302.40    2627.86    1806.28    5937.37    2261.77
#>   9     1902.16     829.78     738.38     776.27     563.45    2149.01
#>   10    1182.49    1743.94    1469.08    1475.94    1787.19    2029.21
#>     year
#> age  2011       2012       2013       2014       2015       2016      
#>   1   163502.68   98149.48  129467.17  208745.31  127812.76  128907.81
#>   2   175630.70  205693.78  177076.40  209915.05  226503.32  113194.28
#>   3   125474.80  184253.10  217019.50  167118.70  179280.50  195889.00
#>   4    92002.60   89537.10  119932.30  140221.90  114154.30  125580.90
#>   5    58469.50   56101.56   54291.80   69447.80   78925.40   65419.30
#>   6    16829.21   33838.48   30423.65   27804.45   36150.81   43845.72
#>   7     7720.61    8357.86   15653.39   15042.83   15185.55   19751.87
#>   8     5236.21    3640.73    4310.38    8559.20    7800.51    6880.26
#>   9      987.75    2763.26    2123.34    2339.97    3873.36    3037.44
#>   10    3674.02    4098.71    5958.17    6821.41    7393.56    8702.57
#>     year
#> age  2017      
#>   1   115058.24
#>   2   143659.04
#>   3   110813.00
#>   4   131782.60
#>   5    75562.10
#>   6    38451.76
#>   7    21449.14
#>   8     7873.59
#>   9     2567.24
#>   10    8882.65
#> 
#> units:  1000 
catch.wt(fca)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  1957     1958     1959     1960     1961     1962     1963     1964    
#>   1  0.043999 0.047000 0.051000 0.045000 0.044000 0.042000 0.048000 0.032000
#>   2  0.109801 0.105531 0.119977 0.115036 0.101425 0.099279 0.109657 0.125989
#>   3  0.194296 0.189472 0.191506 0.203724 0.180139 0.181375 0.174760 0.203920
#>   4  0.256615 0.256336 0.260406 0.287076 0.301163 0.273064 0.304230 0.271018
#>   5  0.349100 0.329376 0.345271 0.377377 0.402368 0.396960 0.391876 0.378894
#>   6  0.455454 0.452345 0.471905 0.479940 0.506119 0.538035 0.531254 0.484717
#>   7  0.532988 0.512971 0.591979 0.600988 0.603987 0.569980 0.623985 0.627988
#>   8  0.588982 0.614964 0.622952 0.682962 0.670966 0.691974 0.666969 0.699965
#>   9  0.395964 0.664949 0.749944 0.723929 0.811933 0.776924 0.714950 0.736908
#>   10 0.997959 0.991965 0.999970 1.093964 1.070964 1.126954 1.027970 1.004970
#>     year
#> age  1965     1966     1967     1968     1969     1970     1971     1972    
#>   1  0.038000 0.038000 0.035999 0.060000 0.052007 0.049160 0.057058 0.066877
#>   2  0.075866 0.104304 0.110549 0.116347 0.174246 0.131116 0.159854 0.207158
#>   3  0.214406 0.148098 0.189574 0.223414 0.271611 0.268303 0.276554 0.289992
#>   4  0.313280 0.314817 0.234828 0.274537 0.283588 0.352160 0.388422 0.407373
#>   5  0.381437 0.427637 0.422016 0.339914 0.356021 0.393751 0.444319 0.486404
#>   6  0.468814 0.482866 0.543256 0.516098 0.408318 0.440916 0.512235 0.539756
#>   7  0.538992 0.558990 0.596989 0.589983 0.572974 0.498998 0.541984 0.607979
#>   8  0.662961 0.623983 0.661974 0.595971 0.654967 0.671964 0.606994 0.645971
#>   9  0.725914 0.689948 0.737946 0.685915 0.657962 0.743909 0.698947 0.673982
#>   10 0.886975 0.932969 0.977957 0.910967 0.892974 0.891969 0.890974 0.938971
#>     year
#> age  1973     1974     1975     1976     1977     1978     1979     1980    
#>   1  0.045283 0.056495 0.068767 0.087787 0.071297 0.069846 0.066755 0.055802
#>   2  0.205176 0.120785 0.151806 0.180926 0.217858 0.190274 0.190466 0.196997
#>   3  0.333995 0.343206 0.204961 0.261605 0.245025 0.315341 0.295421 0.342960
#>   4  0.403174 0.403747 0.392940 0.347254 0.318061 0.364251 0.337713 0.399370
#>   5  0.478238 0.472376 0.492452 0.509426 0.395996 0.432299 0.426454 0.471181
#>   6  0.537635 0.552275 0.585128 0.591583 0.551315 0.486444 0.472122 0.542439
#>   7  0.604968 0.608978 0.635984 0.640979 0.646945 0.608967 0.549993 0.587979
#>   8  0.626970 0.692947 0.702944 0.704946 0.720928 0.686894 0.674948 0.661962
#>   9  0.676947 0.706939 0.782910 0.740877 0.714912 0.775834 0.795858 0.771797
#>   10 0.841983 0.925980 1.018976 0.979962 0.977947 0.949934 0.959948 1.012908
#>     year
#> age  1981     1982     1983     1984     1985     1986     1987     1988    
#>   1  0.048365 0.055830 0.051689 0.053062 0.054057 0.049257 0.043000 0.043000
#>   2  0.183437 0.152001 0.152694 0.149910 0.169134 0.141397 0.113402 0.101804
#>   3  0.326828 0.308248 0.274762 0.265457 0.267967 0.275409 0.218606 0.197485
#>   4  0.414729 0.420607 0.379126 0.321042 0.333496 0.310525 0.343365 0.275873
#>   5  0.501685 0.511917 0.506954 0.470452 0.443875 0.402396 0.375422 0.415128
#>   6  0.555582 0.605900 0.601997 0.588348 0.561543 0.471807 0.470664 0.476689
#>   7  0.603969 0.663973 0.676959 0.676968 0.666955 0.667978 0.573979 0.589985
#>   8  0.641961 0.711936 0.770909 0.725935 0.729903 0.749920 0.727922 0.679945
#>   9  0.724928 0.737919 0.814865 0.838846 0.806857 0.855849 0.834822 0.807882
#>   10 1.006917 0.983944 0.983941 1.035921 1.020912 1.013928 0.992912 1.016921
#>     year
#> age  1989     1990     1991     1992     1993     1994     1995     1996    
#>   1  0.047111 0.053316 0.056328 0.054789 0.062608 0.063660 0.071146 0.053846
#>   2  0.117643 0.129659 0.149493 0.145431 0.160205 0.179072 0.182859 0.140176
#>   3  0.215080 0.210649 0.210932 0.225705 0.249957 0.256510 0.282589 0.267965
#>   4  0.291297 0.290494 0.272753 0.275225 0.303272 0.331360 0.333654 0.336259
#>   5  0.364309 0.359634 0.349364 0.324154 0.339674 0.372247 0.372883 0.412679
#>   6  0.512269 0.439763 0.449276 0.410348 0.406913 0.416102 0.419083 0.463592
#>   7  0.590975 0.585968 0.525990 0.529995 0.511991 0.490993 0.473992 0.489980
#>   8  0.667958 0.689926 0.665960 0.606975 0.629973 0.609971 0.592967 0.552953
#>   9  0.784831 0.760900 0.742884 0.718891 0.719903 0.730933 0.733853 0.711873
#>   10 0.939919 1.009880 0.923918 0.890936 0.855927 0.905917 0.905884 0.857914
#>     year
#> age  1997     1998     1999     2000     2001     2002     2003     2004    
#>   1  0.045137 0.047196 0.053825 0.063148 0.090053 0.056714 0.065718 0.054372
#>   2  0.128986 0.093865 0.103245 0.123154 0.135808 0.130844 0.123961 0.124949
#>   3  0.220379 0.208050 0.198554 0.209624 0.196628 0.221728 0.226407 0.220420
#>   4  0.353209 0.298684 0.266515 0.274163 0.233921 0.285107 0.283692 0.296929
#>   5  0.408398 0.449011 0.413306 0.371930 0.303216 0.326126 0.335699 0.375498
#>   6  0.472671 0.544009 0.414278 0.452321 0.409800 0.426357 0.385257 0.421241
#>   7  0.540972 0.612968 0.537933 0.564982 0.576465 0.469180 0.419098 0.505507
#>   8  0.573943 0.672883 0.636879 0.600437 0.700436 0.643572 0.634464 0.559824
#>   9  0.615899 0.686827 0.747766 0.751606 0.795388 0.759522 0.762447 0.796136
#>   10 0.911896 0.898866 0.803876 0.887791 0.798831 0.903724 0.856513 0.871236
#>     year
#> age  2005     2006     2007     2008     2009     2010     2011     2012    
#>   1  0.067422 0.060317 0.058775 0.056569 0.061207 0.061822 0.047647 0.052327
#>   2  0.117393 0.139368 0.113666 0.123499 0.124699 0.131690 0.114740 0.096187
#>   3  0.212930 0.211812 0.222959 0.247632 0.232112 0.226039 0.211950 0.195563
#>   4  0.297902 0.293315 0.310085 0.324622 0.326594 0.311303 0.277422 0.294436
#>   5  0.351575 0.374795 0.351229 0.389383 0.399101 0.395925 0.369451 0.348424
#>   6  0.347476 0.383401 0.375268 0.436803 0.465827 0.442379 0.453226 0.425402
#>   7  0.453163 0.428475 0.490651 0.368272 0.518436 0.462564 0.595436 0.509182
#>   8  0.554110 0.456878 0.357361 0.468685 0.440965 0.573445 0.444870 0.557246
#>   9  0.616866 0.530799 0.586515 0.639565 0.667432 0.681865 0.555780 0.557920
#>   10 0.909298 0.747735 0.631758 0.637659 0.791518 0.648744 0.803854 0.679823
#>     year
#> age  2013     2014     2015     2016     2017    
#>   1  0.050915 0.025005 0.026000 0.048000 0.051034
#>   2  0.091130 0.093435 0.080271 0.083793 0.085895
#>   3  0.178565 0.171657 0.161845 0.157402 0.158552
#>   4  0.273559 0.253256 0.257901 0.243039 0.218409
#>   5  0.344682 0.317605 0.325621 0.298878 0.314393
#>   6  0.409448 0.396089 0.393707 0.352469 0.386271
#>   7  0.490250 0.472624 0.461250 0.422309 0.437717
#>   8  0.599009 0.541619 0.480815 0.465186 0.531501
#>   9  0.606877 0.627925 0.581978 0.555971 0.641979
#>   10 0.679910 0.649906 0.599938 0.683933 0.734886
#> 
#> units:  kg