FLCatches constructor

Make an FLCatches object.

This is class is used inside FLFishery to store the catches of all species caught by that fleet. It is not meant to be used directly.

FLCatches(object, ...)

# S4 method for list
FLCatches(object)

# S4 method for missing
FLCatches(object, ...)

Arguments

object

Either a list of FLCatch objects or missing

...

Other things

Value

An FLCatches object

Details

Make an FLCatches object.

Validity

Length of dimensions 2:5

All elements must be of class FLCatch

Length of dimensions 2, 4 and 5

All FLQuant slots must share dimensions 2, 4 and 5 (year, season and area).

quant must 'age'

The 1st dimension in elements must be 'age'.

You can inspect the class validity function by using getValidity(getClassDef('FLCatches'))

Accessors

Elements in the classes can be extracted and replaced using the list subset operators,'[', '[<-', '[[' and '[[<-'.

The values passed for replacement need to be of the class FLCatch.

Constructor

A construction method exists for this class that can take named arguments for any of its elements.

Methods

Methods exist for various operations with elements stored in the class:

plot

Standard plot for the FLCatches class.

See also

FLComp

FLCatch, FLFishery