FLFisheries constructor

Make an FLFisheries object.

This is a containuer class for FLFishery objects.

FLFisheries(object, ...)

# S4 method for FLFisheries
landings(object)

# S4 method for FLFisheries
discards(object)

# S4 method for FLFisheries
catch(object)

# S4 method for FLFisheries
lrevenue(object)

# S4 method for FLFisheries
revenue(object)

# S4 method for FLFisheries
cost(object)

# S4 method for FLFisheries
ccost(object)

# S4 method for FLFisheries
vcost(object)

# S4 method for FLFisheries
fcost(object)

# S4 method for FLFisheries
profit(object)

# S4 method for list
FLFisheries(object, desc = character(1))

# S4 method for missing
FLFisheries(object, ...)

Arguments

object
Either a list of FLFishery objects or missing

...
Other things

desc
Description

Value

An FLFisheries object

Details

Make an FLFisheries object.

Accessors

Elements in the classes can be extracted and replaced using the list subset operators,'[', '[<-', '[[' and '[[<-'.

The values passed for replacement need to be of the class FLFishery.

Constructor

A construction method exists for this class that can take named arguments for any of its elements.

See also