An internal class for homogeneous fishing fleets

The same as the FLFishery class but all predictModel slots have been turned into FLQuant objects. The class is used for passing to C++ FLR objects.

Details

What have you done what's in your mind what do you need Where shall we go to let it out What have you seen, we don't know where you've been Life so often blows your candle out Believe in what is right, what's right for you tonight Who knows what the fuck it's all about

Slots

FLFisherycpp objects inherit from FLCatches woth a number of slots added.

.Data

The list of FLCatch object with catch data per stock, (FLCatches).

name

Name of the object, e.g. species or stock code, (character).

desc

Description of the data contents and origin, (character).

range

Ranges of age and years, plusgroup, (numeric).

capacity

Number of boats in the fleet, (FLQuant).

effort

Mean effort per boat applied by the fleet, (FLQuant).

hperiod

Start and end of fishing within each time step, as proportions. An FLQuant object with dimnames `quant=c('start', 'end')` in the first dimension.

vcost

Variable costs per unit of effort, (FLQuant).

fcost

Variable costs per unit of effort, (FLQuant).

orevenue

Revenues obtained from sources other than landings, (FLQuant).

crewshare

Formula, parameter values and inputs to calculate the crew costs, (FLQuant).

Validity

You can inspect the class validity function by using getValidity(getClassDef('FLFishery'))

Accessors

All slots in the class have accessor and replacement methods defined that allow retrieving and substituting individual slots.

The values passed for replacement need to be of the class of that slot. A numeric vector can also be used when replacing FLQuant slots, and the vector will be used to substitute the values in the slot, but not its other attributes.

Constructor

A construction method exists for this class that can take named arguments for any of its slots. All slots are then created to match the requirements of the class validity. If an unnamed FLQuant object is provided, this is used for sizing but not stored in any slot.

Methods

Methods exist for various calculations based on values stored in the class:

ccost

Calculate the total crew costs by evaluating the formula in crewshare.

cost

Total costs, calculated.

lrevenue

.

revenue

.

profit

.

landings

.

discards

.

catch

.

catch.n

.

catch.wt

.

harvest

.

See also

Examples

data(ple4)
FLFishery(PLE=as(ple4, 'FLCatch'))

#> An object of class "FLFishery"
#> [[1]]
#> An object of class "FLCatch"
#> Slot "landings.n":
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  1957   1958   1959   1960   1961   1962   1963   1964   1965   1966  
#>   1       0      0      0      0      0      0      0      0      0      0
#>   2    4315   7129  16556   5959   2264   2147   4340  14708   9858   4144
#>   3   59818  22205  30427  61876  33392  35876  21471  40486  42202  65009
#>   4   44718  62047  25489  51022  67906  66779  76926  64735  53188  51488
#>   5   31771  34112  41099  21321  32699  50060  54364  57408  43674  36667
#>   6    8885  19594  22936  27329  12759  20628  31799  37091  30151  27370
#>   7   11029   8178  13873  14186  14680   9060  12848  15819  18361  16500
#>   8    9028   8000   6408   9013   9748   9035   6833   6595   8554  10784
#>   9    4973   6110   6596   5087   5996   5257   7047   3980   4213   6467
#>   10  10859  13148  16180  15153  14660  12801  16592  16886  17587  14928
#>     year
#> age  1967   1968   1969   1970   1971   1972   1973   1974   1975   1976  
#>   1       0      0      3     76     19   2233   1268   2223    981   2820
#>   2    5982   9474  15017  17294  29591  36528  31733  23120  28124  33643
#>   3   30304  40698  45187  51174  48282  62199  59099  55548  61623  77649
#>   4  112917  38140  36084  56153  33475  52906  73065  42125  31262  96398
#>   5   41383 123619  35585  40686  26059  23043  42255  41075  25419  13779
#>   6   22053  17139 102014  35074  22903  16998  13817  19666  21188   9904
#>   7   16175  10341  10410  78886  16913  14380   8885   8005  11873   9120
#>   8    8004  10102   6086   6311  29730  10903   9848   6321   5923   6391
#>   9    6728   3925   8192   4185   6414  18585   6084   5568   4106   2947
#>   10  11175  13365  16092  14840  16910  15651  23978  21980  19695  12552
#>     year
#> age  1977   1978   1979   1980   1981   1982   1983   1984   1985   1986  
#>   1    3220   1143   1318    979    253   3334   1214    108    121   1674
#>   2   56969  60578  58031  64904 100927  47776 119695  63252  73552  67125
#>   3   43289  62343 118863 133741 122296 209007 115034 274209 144316 163717
#>   4   66013  54341  48962  77523  57604  69544  99076  53549 185203  93801
#>   5   83705  50102  47886  24974  35745  28655  29359  37468  32520  84479
#>   6    9142  35510  39932  17982  12414  16726  12906  13661  15544  24049
#>   7    5912   5940  24228  13761   9564   7589   8216   6465   6871   9299
#>   8    5022   3352   4161   8458   8092   5470   4193   5544   3650   4490
#>   9    4061   2419   2807   1864   4874   4482   3013   2720   2698   2733
#>   10   9191   7468   9288   5377   5903   8653   8287   6565   5798   6950
#>     year
#> age  1987   1988   1989   1990   1991   1992   1993   1994   1995   1996  
#>   1       0      0   1261   1550   1461   3410   3461   1394   7751   1104
#>   2   85123  15146  46757  32533  43266  43954  53949  45148  36575  42496
#>   3  115951 250675 105929  97766  83603  85120  98375 101617  81398  64382
#>   4  111239  74335 231414 110997 116155  72494  72286  80236  78370  46359
#>   5   64758  47380  52909 159814  72961  72703  51405  38542  36499  32130
#>   6   34728  25091  19247  26757  77557  33406  29001  20388  17953  14460
#>   7   11452  16774  10567   8129  14910  29547  13472  15323   9772  10605
#>   8    4341   5381   7561   4216   5233   6970  11272   6399   4366   4528
#>   9    2154   3162   2120   3451   3141   3200   3645   5368   2336   2624
#>   10   5478   6233   5580   3808   5591   6928   5883   5433   3753   4892
#>     year
#> age  1997   1998   1999   2000   2001   2002   2003   2004   2005   2006  
#>   1     892    196    549   2634   4509   1233    694    543   2937    355
#>   2   42855  30401   8689  15819  35886  15596  42594  10317  16685  18987
#>   3   86948  68920 155971  39550  52480  58262  47802 102332  26069  67465
#>   4   43669  56329  39857 164330  48238  48361  48894  35165  82278  25254
#>   5   22541  16713  24112  14993  89949  36551  27126  20527  17039  42525
#>   6   13518   6432   6829   9343   6836  37877  15999  11293   9533   6555
#>   7    6362   4986   2783   2130   4418   4644  17069   4787   5332   4967
#>   8    3632   2506   2246   1030   1127   1788   1608   4555   2614   2053
#>   9    2179   1761   1521    940    637    742    650    412   2223   1235
#>   10   4181   3119   3093   2097   2309   1586    859    540    613   1319
#>     year
#> age  2007   2008  
#>   1    1286    380
#>   2   19205  10970
#>   3   37309  42865
#>   4   47053  37970
#>   5   14971  29476
#>   6   17142   5700
#>   7    2459   6752
#>   8    1856    912
#>   9     543    673
#>   10   1259    896
#> 
#> units:  10^3 
#> 
#> Slot "landings.wt":
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  1957    1958    1959    1960    1961    1962    1963    1964    1965   
#>   1  0.00000 0.00000 0.00000 0.00000 0.00000 0.00000 0.00000 0.00000 0.00000
#>   2  0.18323 0.21055 0.22273 0.20133 0.19397 0.20394 0.25788 0.25187 0.24338
#>   3  0.22321 0.23500 0.25134 0.23757 0.23662 0.23970 0.29152 0.27505 0.28360
#>   4  0.28651 0.27541 0.29936 0.29092 0.30670 0.28996 0.32515 0.31434 0.32282
#>   5  0.39200 0.35836 0.36986 0.38857 0.41841 0.38662 0.40670 0.39091 0.38719
#>   6  0.50638 0.48171 0.48327 0.48823 0.51692 0.52290 0.54328 0.49065 0.47368
#>   7  0.59189 0.54551 0.60485 0.60500 0.61340 0.55093 0.63604 0.63270 0.54206
#>   8  0.65408 0.65397 0.63653 0.68755 0.68144 0.66885 0.67987 0.70524 0.66677
#>   9  0.43975 0.70714 0.76628 0.72882 0.82463 0.75100 0.72879 0.74252 0.73013
#>   10 1.10829 1.05490 1.02131 1.10101 1.08787 1.08964 1.04798 1.01246 0.89160
#>     year
#> age  1966    1967    1968    1969    1970    1971    1972    1973    1974   
#>   1  0.00000 0.00000 0.00000 0.22963 0.30693 0.26447 0.25296 0.28581 0.29553
#>   2  0.23623 0.23660 0.27478 0.31111 0.27867 0.32852 0.30438 0.33204 0.32249
#>   3  0.27492 0.28453 0.30668 0.32804 0.30985 0.36777 0.36196 0.36146 0.36708
#>   4  0.35434 0.32838 0.34064 0.35238 0.34688 0.41633 0.44012 0.42556 0.41996
#>   5  0.44394 0.43343 0.37666 0.37989 0.40827 0.46282 0.50696 0.51067 0.49358
#>   6  0.49282 0.55784 0.53206 0.43597 0.43165 0.53100 0.55632 0.56636 0.57446
#>   7  0.56919 0.60884 0.60719 0.60634 0.48622 0.55993 0.62521 0.63571 0.63150
#>   8  0.63537 0.67512 0.61336 0.69311 0.65478 0.62708 0.66429 0.65883 0.71860
#>   9  0.70257 0.75263 0.70599 0.69629 0.72494 0.72212 0.69308 0.71137 0.73312
#>   10 0.95022 0.99752 0.93747 0.94534 0.86876 0.92002 0.96536 0.88440 0.95975
#>     year
#> age  1975    1976    1977    1978    1979    1980    1981    1982    1983   
#>   1  0.26454 0.27173 0.25441 0.23528 0.23460 0.24123 0.24115 0.28074 0.19875
#>   2  0.31872 0.30249 0.32352 0.30374 0.31047 0.28988 0.27880 0.26364 0.24844
#>   3  0.35059 0.34658 0.35357 0.35581 0.34840 0.34867 0.33476 0.31294 0.29813
#>   4  0.44621 0.38452 0.38061 0.38281 0.38734 0.40644 0.42329 0.42665 0.38061
#>   5  0.52589 0.52602 0.41868 0.42234 0.42827 0.47941 0.51384 0.51721 0.51179
#>   6  0.62363 0.60907 0.55690 0.47345 0.47319 0.55239 0.56777 0.61180 0.60024
#>   7  0.67569 0.65727 0.64805 0.58723 0.54906 0.59597 0.61458 0.66815 0.67278
#>   8  0.74687 0.72289 0.72217 0.66244 0.67385 0.67098 0.65324 0.71645 0.76620
#>   9  0.83186 0.75981 0.71616 0.74826 0.79464 0.78247 0.73770 0.74261 0.80992
#>   10 1.08231 1.00505 0.97968 0.91581 0.95862 1.02671 1.02481 0.98991 0.97766
#>     year
#> age  1984    1985    1986    1987    1988    1989    1990    1991    1992   
#>   1  0.22936 0.24204 0.21828 0.21825 0.21763 0.23256 0.26683 0.21869 0.24643
#>   2  0.25889 0.25869 0.26569 0.24590 0.25013 0.27592 0.28061 0.27554 0.25821
#>   3  0.27858 0.28417 0.30026 0.29626 0.27377 0.30450 0.29341 0.28324 0.28472
#>   4  0.36914 0.33023 0.34273 0.34663 0.34664 0.32717 0.31212 0.29480 0.31221
#>   5  0.48333 0.45272 0.41977 0.39699 0.44610 0.38629 0.36036 0.35165 0.33480
#>   6  0.60343 0.56540 0.48200 0.49772 0.50420 0.52524 0.44011 0.43835 0.41727
#>   7  0.67332 0.66437 0.66670 0.57574 0.59874 0.59422 0.58780 0.50868 0.52134
#>   8  0.71368 0.71435 0.74176 0.71893 0.68835 0.66024 0.68134 0.64645 0.59399
#>   9  0.82393 0.78784 0.84251 0.81868 0.80062 0.78047 0.74928 0.71967 0.70199
#>   10 1.01857 1.00080 1.00069 0.97782 0.99898 0.92899 0.98855 0.88688 0.87477
#>     year
#> age  1993    1994    1995    1996    1997    1998    1999    2000    2001   
#>   1  0.24320 0.22299 0.27022 0.23630 0.20633 0.14959 0.24158 0.22142 0.23565
#>   2  0.26664 0.25610 0.27519 0.27568 0.26883 0.25588 0.24947 0.25929 0.26437
#>   3  0.28227 0.27849 0.29903 0.30227 0.31049 0.30509 0.27609 0.27580 0.28912
#>   4  0.31840 0.33010 0.33579 0.34952 0.36108 0.38776 0.35005 0.30494 0.30595
#>   5  0.34770 0.38658 0.39937 0.41352 0.45333 0.48913 0.44865 0.41953 0.36140
#>   6  0.41314 0.43721 0.45103 0.47850 0.51979 0.59739 0.53937 0.48557 0.47725
#>   7  0.50593 0.48882 0.52455 0.49130 0.59816 0.62297 0.62121 0.66426 0.58616
#>   8  0.61630 0.59496 0.60701 0.57992 0.61106 0.68399 0.67248 0.68951 0.70102
#>   9  0.70420 0.71279 0.72921 0.70890 0.67752 0.68891 0.74150 0.72933 0.78716
#>   10 0.83590 0.88280 0.90219 0.84440 0.91674 0.89965 0.80164 0.86173 0.79266
#>     year
#> age  2002    2003    2004    2005    2006    2007    2008   
#>   1  0.23208 0.22729 0.21192 0.26739 0.25667 0.26158 0.24721
#>   2  0.25852 0.24787 0.24460 0.26152 0.27189 0.26652 0.26508
#>   3  0.28300 0.28118 0.28025 0.27719 0.28914 0.30304 0.30579
#>   4  0.30943 0.31938 0.32481 0.32714 0.33784 0.34549 0.34253
#>   5  0.34077 0.36347 0.39413 0.38493 0.39871 0.37806 0.40309
#>   6  0.43575 0.40560 0.43275 0.42704 0.40885 0.45209 0.45273
#>   7  0.50038 0.47712 0.50504 0.46328 0.47480 0.53896 0.53811
#>   8  0.67762 0.64073 0.55158 0.54458 0.48900 0.48072 0.72576
#>   9  0.74519 0.75046 0.78924 0.60335 0.53262 0.59029 0.64037
#>   10 0.88125 0.83724 0.86088 0.88778 0.75508 0.61872 0.63675
#> 
#> units:  kg 
#> 
#> Slot "discards.n":
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  1957    1958    1959    1960    1961    1962    1963    1964    1965   
#>   1    32356   66199  116086   73939   75578   51265   90913   66035   43708
#>   2    45596   73552  127771  167893  144609  181321  136183  153274  426021
#>   3     9220   23655   46402   44948   89014   87599  129778   64156   59262
#>   4      909    2572   11407     997     538   21716    9964   33825    3404
#>   5      961    2137    4737    1067    1612     799    2112    3011     923
#>   6       25      65     106     519     130     186     188     323     267
#>   7        0       0       0       0       0       0       0       0       0
#>   8        0       0       0       0       0       0       0       0       0
#>   9        0       0       0       0       0       0       0       0       0
#>   10       0       0       0       0       0       0       0       0       0
#>     year
#> age  1966    1967    1968    1969    1970    1971    1972    1973    1974   
#>   1    38496   20199   73971   85192  123569   69337   70002  132352  211139
#>   2   163125  133545   72192   67378  152480   96968   55470   49815  308411
#>   3   349358   87532   46339   16747   27747   42354   33899    4008    3652
#>   4    14399  152496   26530   19334    1287    2675    5714     673     285
#>   5     1402     623   22436     773    5061     426     567    1289     611
#>   6      125     260      58    2024     161      81      73      67     109
#>   7        0       0       0       0       0       0       0       0       0
#>   8        0       0       0       0       0       0       0       0       0
#>   9        0       0       0       0       0       0       0       0       0
#>   10       0       0       0       0       0       0       0       0       0
#>     year
#> age  1975    1976    1977    1978    1979    1980    1981    1982    1983   
#>   1   244969  183879  256628  226872  293166  226371  134142  411307  261400
#>   2   280130  140921  103696  154113  215084  122561  193241  204572  436331
#>   3   190536   71054   79317   27257   57578     932    1850    4624   30716
#>   4     4807   18013   33552   10775   18382     687     373    1109    2235
#>   5      253     174    9317    1244     589     193     431     216     804
#>   6      123      41     129     570     310      86      55      98      72
#>   7        0       0       0       0       0       0       0       0       0
#>   8        0       0       0       0       0       0       0       0       0
#>   9        0       0       0       0       0       0       0       0       0
#>   10       0       0       0       0       0       0       0       0       0
#>     year
#> age  1984    1985    1986    1987    1988    1989    1990    1991    1992   
#>   1   310675  405385 1117345  361519  348597  213291  145314  183126  138755
#>   2   313490  229208  490965 1374202  608109  485845  279298  301575  219619
#>   3    52651   35566   48510  180969  459385  193176  168674  141567   94581
#>   4    24529    2221   26470    1427   61167   85758   28102   40739   34348
#>   5     1492     200    1451    1348     882    7224    5011    5528    4307
#>   6       69      78     146     248     177     115     177     939     880
#>   7        0       0       0       0       0       0       0       0       0
#>   8        0       0       0       0       0       0       0       0       0
#>   9        0       0       0       0       0       0       0       0       0
#>   10       0       0       0       0       0       0       0       0       0
#>     year
#> age  1993    1994    1995    1996    1997    1998    1999    2000    2001   
#>   1    96371   62122  118863  111250  128653  104538  127321  103422   30295
#>   2   154083   95703   82676  331065  510918  646250  208401  171828  352922
#>   3    48088   35703   15753   27606  193828  191631  231769   52354  187964
#>   4    11966    1038     860    3930     588   53354   54869   65871   75433
#>   5     1635     822     663     451     271     297     278    1632   54602
#>   6      216     144     120     116     108      33      58     312     212
#>   7        0       0       0       0       0       0       0     269      46
#>   8        0       0       0       0       0       0       0     168       1
#>   9        0       0       0       0       0       0       0       0       0
#>   10       0       0       0       0       0       0       0       0       0
#>     year
#> age  2002    2003    2004    2005    2006    2007    2008   
#>   1   309770   67615  232854   93541  220473   77312  135406
#>   2   178382  518194  180824  325997  225077  205140  252629
#>   3    78012   53760  119011   46007  110215   69312   37393
#>   4    13378   19942    8073   20842   10656   10735    6237
#>   5     2130    4213    1674    4707    3000    1437    2282
#>   6      761     455     357    6077     410    7151     517
#>   7      119    5752      43     157     754     204    8882
#>   8        1       1       1       1     194    1649     891
#>   9        0       0       0       0       0       0       0
#>   10       0       0       0       0       0       0       0
#> 
#> units:  10^3 
#> 
#> Slot "discards.wt":
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  1957  1958  1959  1960  1961  1962  1963  1964  1965  1966  1967  1968 
#>   1  0.046 0.049 0.053 0.047 0.046 0.044 0.049 0.034 0.040 0.040 0.038 0.062
#>   2  0.102 0.095 0.106 0.110 0.098 0.097 0.103 0.112 0.071 0.100 0.104 0.095
#>   3  0.147 0.158 0.155 0.159 0.160 0.155 0.156 0.160 0.165 0.126 0.157 0.156
#>   4  0.180 0.186 0.185 0.186 0.192 0.192 0.186 0.191 0.191 0.192 0.169 0.190
#>   5  0.204 0.198 0.193 0.199 0.202 0.211 0.203 0.202 0.210 0.203 0.211 0.189
#>   6  0.231 0.244 0.231 0.210 0.212 0.219 0.231 0.220 0.220 0.231 0.219 0.244
#>   7  0.244 0.244 0.000 0.231 0.211 0.220 0.220 0.231 0.220 0.220 0.231 0.212
#>   8  0.231 0.000 0.000 0.000 0.244 0.220 0.231 0.231 0.000 0.231 0.244 0.000
#>   9  0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#>   10 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#>     year
#> age  1969  1970  1971  1972  1973  1974  1975  1976  1977  1978  1979  1980 
#>   1  0.055 0.051 0.059 0.063 0.045 0.057 0.070 0.088 0.071 0.071 0.069 0.057
#>   2  0.144 0.113 0.109 0.145 0.128 0.105 0.134 0.151 0.157 0.141 0.155 0.147
#>   3  0.161 0.178 0.182 0.173 0.178 0.173 0.162 0.175 0.185 0.196 0.184 0.190
#>   4  0.185 0.187 0.198 0.210 0.193 0.210 0.204 0.193 0.193 0.203 0.202 0.211
#>   5  0.205 0.192 0.211 0.205 0.204 0.212 0.220 0.219 0.196 0.205 0.219 0.220
#>   6  0.210 0.000 0.000 0.244 0.231 0.231 0.244 0.244 0.211 0.211 0.231 0.000
#>   7  0.244 0.212 0.000 0.000 0.244 0.244 0.244 0.244 0.000 0.220 0.220 0.244
#>   8  0.231 0.231 0.231 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.244 0.000
#>   9  0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#>   10 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#>     year
#> age  1981  1982  1983  1984  1985  1986  1987  1988  1989  1990  1991  1992 
#>   1  0.050 0.056 0.054 0.055 0.056 0.051 0.044 0.044 0.047 0.054 0.058 0.055
#>   2  0.133 0.125 0.124 0.125 0.138 0.123 0.104 0.098 0.102 0.113 0.130 0.124
#>   3  0.178 0.182 0.180 0.172 0.176 0.180 0.165 0.154 0.163 0.159 0.162 0.167
#>   4  0.210 0.199 0.203 0.210 0.195 0.192 0.203 0.184 0.180 0.185 0.185 0.186
#>   5  0.219 0.231 0.205 0.203 0.231 0.211 0.211 0.211 0.192 0.205 0.199 0.200
#>   6  0.244 0.231 0.244 0.000 0.244 0.244 0.231 0.231 0.244 0.231 0.220 0.210
#>   7  0.000 0.244 0.244 0.244 0.000 0.231 0.000 0.000 0.000 0.000 0.220 0.220
#>   8  0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.231 0.244
#>   9  0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#>   10 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#>     year
#> age  1993  1994  1995  1996  1997  1998  1999  2000  2001  2002  2003  2004 
#>   1  0.059 0.062 0.060 0.053 0.042 0.049 0.055 0.061 0.070 0.058 0.068 0.057
#>   2  0.120 0.141 0.140 0.122 0.118 0.086 0.096 0.110 0.122 0.119 0.114 0.117
#>   3  0.171 0.175 0.185 0.178 0.160 0.168 0.144 0.152 0.167 0.171 0.174 0.167
#>   4  0.196 0.192 0.199 0.203 0.202 0.197 0.193 0.173 0.177 0.191 0.184 0.192
#>   5  0.205 0.211 0.212 0.220 0.220 0.212 0.211 0.231 0.195 0.196 0.198 0.196
#>   6  0.231 0.231 0.231 0.231 0.244 0.000 0.244 0.000 0.231 0.211 0.204 0.211
#>   7  0.231 0.244 0.231 0.000 0.000 0.244 0.000 0.198 0.000 0.000 0.220 0.199
#>   8  0.000 0.220 0.244 0.244 0.000 0.000 0.000 0.000 0.231 0.000 0.000 0.000
#>   9  0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#>   10 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#>     year
#> age  2005  2006  2007  2008 
#>   1  0.063 0.062 0.057 0.058
#>   2  0.109 0.128 0.097 0.116
#>   3  0.172 0.162 0.177 0.165
#>   4  0.178 0.193 0.178 0.187
#>   5  0.220 0.197 0.192 0.188
#>   6  0.204 0.199 0.197 0.231
#>   7  0.219 0.210 0.220 0.220
#>   8  0.231 0.211 0.197 0.190
#>   9  0.000 0.000 0.000 0.000
#>   10 0.000 0.000 0.000 0.000
#> 
#> units:  kg 
#> 
#> Slot "catch.sel":
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  1957     1958     1959     1960     1961     1962     1963     1964    
#>   1  0.222728 0.280897 0.370201 0.258776 0.221695 0.221045 0.315109 0.058519
#>   2  0.659688 0.668381 0.753871 0.760049 0.661758 0.734272 0.767065 0.739437
#>   3  0.736912 0.808277 0.862097 0.844232 0.788356 0.860093 0.882130 0.829838
#>   4  0.877310 0.957570 0.914442 0.917216 0.816549 0.917887 0.914326 0.867953
#>   5  1.000000 1.000000 1.000000 0.874585 0.954792 1.000000 0.892498 1.000000
#>   6  0.601602 0.859734 0.931396 1.000000 0.755680 0.982255 1.000000 0.902922
#>   7  0.789616 0.716916 0.851377 0.914276 0.825343 0.834814 0.949526 0.746418
#>   8  0.907026 0.777629 0.751830 0.856929 1.000000 0.806621 0.953854 0.721700
#>   9  0.837160 0.864655 0.892389 0.915326 0.873065 0.911088 0.945177 0.850669
#>   10 0.837160 0.864655 0.892389 0.915326 0.873065 0.911088 0.945177 0.850669
#>     year
#> age  1965     1966     1967     1968     1969     1970     1971     1972    
#>   1  0.134673 0.141248 0.107773 0.481788 0.347050 0.375537 0.457613 0.485310
#>   2  0.525106 0.703891 0.697817 0.701320 0.730574 0.732725 0.776246 0.796695
#>   3  0.781112 0.766214 0.802955 0.839330 0.763856 0.828895 0.906900 0.838404
#>   4  0.811483 0.850631 0.809346 0.881040 0.795809 0.850444 0.907903 0.864390
#>   5  0.699096 0.943235 0.944275 0.894249 0.735235 0.778996 0.952675 0.876376
#>   6  1.000000 0.678805 1.000000 0.788863 1.000000 0.849684 0.924018 0.927267
#>   7  0.820111 1.000000 0.614176 1.000000 0.688210 1.000000 1.000000 0.853147
#>   8  0.692536 0.808283 0.863287 0.704677 0.932910 0.439236 0.963015 1.000000
#>   9  0.807165 0.858981 0.848947 0.856162 0.832790 0.786354 0.952481 0.907179
#>   10 0.807165 0.858981 0.848947 0.856162 0.832790 0.786354 0.952481 0.907179
#>     year
#> age  1973     1974     1975     1976     1977     1978     1979     1980    
#>   1  0.207708 0.369931 0.574957 0.643042 0.486061 0.586282 0.603832 0.358401
#>   2  0.723492 0.668329 0.811321 0.786198 0.708809 0.825375 0.900813 0.706829
#>   3  0.794636 0.823042 0.858779 0.822546 0.743544 0.893749 0.941284 1.000000
#>   4  0.993770 0.863969 0.900783 0.834415 0.750740 0.906926 0.953863 0.936522
#>   5  1.000000 1.000000 0.970744 0.739735 1.000000 0.887279 0.965869 0.763226
#>   6  0.757589 0.758267 1.000000 0.836827 0.631240 1.000000 1.000000 0.777915
#>   7  0.710147 0.661040 0.772441 1.000000 0.662678 0.888287 0.995191 0.740423
#>   8  0.880763 0.779195 0.814064 0.873095 0.800885 0.821156 0.855128 0.753924
#>   9  0.871462 0.815353 0.895023 0.859654 0.771915 0.903823 0.958169 0.797360
#>   10 0.871462 0.815353 0.895023 0.859654 0.771915 0.903823 0.958169 0.797360
#>     year
#> age  1981     1982     1983     1984     1985     1986     1987     1988    
#>   1  0.298110 0.348310 0.318660 0.472129 0.373916 0.395506 0.282497 0.318685
#>   2  0.814946 0.741946 0.697874 0.870605 0.673989 0.841717 0.828845 0.874455
#>   3  0.969557 1.000000 0.754776 0.921126 0.704405 0.872683 0.867200 0.925606
#>   4  1.000000 0.971643 1.000000 0.931944 1.000000 0.882959 0.927368 0.922670
#>   5  0.977005 0.853794 0.813361 1.000000 0.668236 1.000000 1.000000 0.963337
#>   6  0.753664 0.747808 0.689904 0.854855 0.712762 0.926774 0.828700 0.954160
#>   7  0.846242 0.689528 0.614547 0.727089 0.726014 0.768823 0.883949 0.906477
#>   8  0.894442 0.768556 0.630704 0.893170 0.650700 0.902412 0.618440 1.000000
#>   9  0.897619 0.809388 0.752602 0.884825 0.754339 0.900003 0.855375 0.953529
#>   10 0.897619 0.809388 0.752602 0.884825 0.754339 0.900003 0.855375 0.953529
#>     year
#> age  1989     1990     1991     1992     1993     1994     1995     1996    
#>   1  0.328472 0.221425 0.286506 0.270984 0.292838 0.183559 0.154846 0.110217
#>   2  0.903127 0.649551 0.727373 0.703453 0.640345 0.548070 0.596052 0.632609
#>   3  0.992000 0.782565 0.792941 0.830948 0.801183 0.682294 0.853896 0.824464
#>   4  1.000000 0.840529 0.833897 0.858507 0.861931 0.807182 1.000000 0.919031
#>   5  0.984879 1.000000 0.900171 1.000000 1.000000 0.726180 0.997779 0.908765
#>   6  0.930233 0.787927 1.000000 0.986312 0.924111 0.746695 0.810329 0.842329
#>   7  0.967137 0.654394 0.773417 0.991667 0.944787 1.000000 0.900981 1.000000
#>   8  0.973090 0.655235 0.685264 0.801136 0.915329 0.897662 0.791592 0.869778
#>   9  0.975104 0.790704 0.842470 0.934271 0.939147 0.842454 0.895580 0.987881
#>   10 0.975104 0.790704 0.842470 0.934271 0.939147 0.842454 0.895580 0.987881
#>     year
#> age  1997     1998     1999     2000     2001     2002     2003     2004    
#>   1  0.069108 0.132655 0.138624 0.165113 0.072423 0.234105 0.137789 0.326064
#>   2  0.830143 0.435360 0.384466 0.478991 0.795448 0.654679 0.635423 0.984274
#>   3  0.986606 0.875312 0.432280 0.446047 1.000000 0.676420 0.644612 0.815228
#>   4  0.855658 1.000000 1.000000 0.834529 0.940822 0.712991 0.652644 0.931983
#>   5  1.000000 0.677230 0.625527 0.872937 0.985432 0.828901 0.733124 0.655907
#>   6  0.854056 0.611298 0.616364 0.963902 0.596135 0.730638 0.776775 1.000000
#>   7  0.772562 0.626528 0.544164 0.779111 0.828368 1.000000 1.000000 0.808121
#>   8  0.781031 0.565308 0.612154 0.818950 0.556687 0.903656 0.845238 0.820781
#>   9  0.966573 0.793744 0.785007 1.000000 0.727000 0.828286 0.740866 0.798919
#>   10 0.966573 0.793744 0.785007 1.000000 0.727000 0.828286 0.740866 0.798919
#>     year
#> age  2005     2006     2007     2008    
#>   1  0.186881 0.389843 0.155420 0.424983
#>   2  0.698415 0.727482 1.000000 1.000000
#>   3  0.702758 0.684503 0.917038 0.867660
#>   4  0.661360 0.655807 0.573448 0.904833
#>   5  0.950915 0.506298 0.813099 0.637138
#>   6  0.790706 0.633165 0.655454 0.676902
#>   7  1.000000 0.539715 0.616001 0.928889
#>   8  0.799421 1.000000 0.861406 0.902320
#>   9  0.712270 0.772548 0.762063 0.344113
#>   10 0.712270 0.772548 0.762063 0.344113
#> 
#> units:   
#> 
#> Slot "price":
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  1957 1958 1959 1960 1961 1962 1963 1964 1965 1966 1967 1968 1969 1970 1971
#>   1  NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>   2  NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>   3  NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>   4  NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>   5  NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>   6  NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>   7  NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>   8  NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>   9  NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>   10 NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>     year
#> age  1972 1973 1974 1975 1976 1977 1978 1979 1980 1981 1982 1983 1984 1985 1986
#>   1  NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>   2  NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>   3  NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>   4  NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>   5  NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>   6  NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>   7  NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>   8  NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>   9  NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>   10 NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>     year
#> age  1987 1988 1989 1990 1991 1992 1993 1994 1995 1996 1997 1998 1999 2000 2001
#>   1  NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>   2  NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>   3  NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>   4  NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>   5  NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>   6  NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>   7  NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>   8  NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>   9  NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>   10 NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>     year
#> age  2002 2003 2004 2005 2006 2007 2008
#>   1  NA   NA   NA   NA   NA   NA   NA  
#>   2  NA   NA   NA   NA   NA   NA   NA  
#>   3  NA   NA   NA   NA   NA   NA   NA  
#>   4  NA   NA   NA   NA   NA   NA   NA  
#>   5  NA   NA   NA   NA   NA   NA   NA  
#>   6  NA   NA   NA   NA   NA   NA   NA  
#>   7  NA   NA   NA   NA   NA   NA   NA  
#>   8  NA   NA   NA   NA   NA   NA   NA  
#>   9  NA   NA   NA   NA   NA   NA   NA  
#>   10 NA   NA   NA   NA   NA   NA   NA  
#> 
#> units:  10^3 
#> 
#> Slot "catch.q":
#> An object of class "FLPar"
#> params
#> alpha  beta 
#>     1     0 
#> units:  NA 
#> 
#> Slot "name":
#> [1] "Plaice in IV"
#> 
#> Slot "desc":
#> [1] "Imported from a VPA file. ( N:\\Projecten\\ICES WG\\Demersale werkgroep WGNSSK\\2009\\stock\\ple-nsea\\final runs\\index.txt ).  Tue Jun 16 06:32:20 2009 + FLAssess: "
#> 
#> Slot "range":
#>       min       max plusgroup   minyear   maxyear 
#>         1        10        10      1957      2008 
#> 
#> 
#> Slot "capacity":
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1957 1958 1959 1960 1961 1962 1963 1964 1965 1966 1967 1968 1969 1970
#>   all NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>      year
#> quant 1971 1972 1973 1974 1975 1976 1977 1978 1979 1980 1981 1982 1983 1984
#>   all NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>      year
#> quant 1985 1986 1987 1988 1989 1990 1991 1992 1993 1994 1995 1996 1997 1998
#>   all NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>      year
#> quant 1999 2000 2001 2002 2003 2004 2005 2006 2007 2008
#>   all NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#> 
#> units:  NA 
#> 
#> Slot "effort":
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1957 1958 1959 1960 1961 1962 1963 1964 1965 1966 1967 1968 1969 1970
#>   all NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>      year
#> quant 1971 1972 1973 1974 1975 1976 1977 1978 1979 1980 1981 1982 1983 1984
#>   all NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>      year
#> quant 1985 1986 1987 1988 1989 1990 1991 1992 1993 1994 1995 1996 1997 1998
#>   all NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>      year
#> quant 1999 2000 2001 2002 2003 2004 2005 2006 2007 2008
#>   all NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#> 
#> units:  NA 
#> 
#> Slot "hperiod":
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>        year
#> quant   1957 1958 1959 1960 1961 1962 1963 1964 1965 1966 1967 1968 1969 1970
#>   start 0    0    0    0    0    0    0    0    0    0    0    0    0    0   
#>   end   1    1    1    1    1    1    1    1    1    1    1    1    1    1   
#>        year
#> quant   1971 1972 1973 1974 1975 1976 1977 1978 1979 1980 1981 1982 1983 1984
#>   start 0    0    0    0    0    0    0    0    0    0    0    0    0    0   
#>   end   1    1    1    1    1    1    1    1    1    1    1    1    1    1   
#>        year
#> quant   1985 1986 1987 1988 1989 1990 1991 1992 1993 1994 1995 1996 1997 1998
#>   start 0    0    0    0    0    0    0    0    0    0    0    0    0    0   
#>   end   1    1    1    1    1    1    1    1    1    1    1    1    1    1   
#>        year
#> quant   1999 2000 2001 2002 2003 2004 2005 2006 2007 2008
#>   start 0    0    0    0    0    0    0    0    0    0   
#>   end   1    1    1    1    1    1    1    1    1    1   
#> 
#> units:  NA 
#> 
#> Slot "vcost":
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1957 1958 1959 1960 1961 1962 1963 1964 1965 1966 1967 1968 1969 1970
#>   all NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>      year
#> quant 1971 1972 1973 1974 1975 1976 1977 1978 1979 1980 1981 1982 1983 1984
#>   all NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>      year
#> quant 1985 1986 1987 1988 1989 1990 1991 1992 1993 1994 1995 1996 1997 1998
#>   all NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>      year
#> quant 1999 2000 2001 2002 2003 2004 2005 2006 2007 2008
#>   all NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#> 
#> units:  NA 
#> 
#> Slot "fcost":
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1957 1958 1959 1960 1961 1962 1963 1964 1965 1966 1967 1968 1969 1970
#>   all NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>      year
#> quant 1971 1972 1973 1974 1975 1976 1977 1978 1979 1980 1981 1982 1983 1984
#>   all NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>      year
#> quant 1985 1986 1987 1988 1989 1990 1991 1992 1993 1994 1995 1996 1997 1998
#>   all NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>      year
#> quant 1999 2000 2001 2002 2003 2004 2005 2006 2007 2008
#>   all NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#> 
#> units:  NA 
#> 
#> Slot "orevenue":
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1957 1958 1959 1960 1961 1962 1963 1964 1965 1966 1967 1968 1969 1970
#>   all NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>      year
#> quant 1971 1972 1973 1974 1975 1976 1977 1978 1979 1980 1981 1982 1983 1984
#>   all NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>      year
#> quant 1985 1986 1987 1988 1989 1990 1991 1992 1993 1994 1995 1996 1997 1998
#>   all NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>      year
#> quant 1999 2000 2001 2002 2003 2004 2005 2006 2007 2008
#>   all NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#> 
#> units:  NA 
#> 
#> Slot "crewshare":
#> An object of class "FLQuants": EMPTY
#> model:  
#> ~fixed * crew + share * lrevenue
#> <environment: namespace:FLFishery>
#> 
#> params:  
#> An object of class "FLPar"
#> params
#> fixed  crew share 
#>     0     1     0 
#> units:  NA 
#> 
#> Slot "name":
#> [1] ""
#> 
#> Slot "desc":
#> [1] ""
#> 
#> Slot "range":
#>     min     max minyear maxyear 
#>       0       0       1       1 
#> 
#> Slot "names":
#> [1] "PLE"
#> 
#> Slot "lock":
#> [1] FALSE
#>