Methods to calculate fishing mortalities

Fishing mortalities and harvest rates can be calculated for a combination of FLBiol and FLFishery/FLFisheries using the harvest() method.

harvests(object, catches, ...)

# S4 method for FLBiol,FLFisheries
harvests(object, catches, fcb = rep(1, length(catches)), units = c("f", "hr"))

# S4 method for FLBiol,FLFishery
harvest(object, catch, fcb = 1)

# S4 method for FLBiol,FLFisheries
harvest(object, catch, fcb = 1)

# S4 method for FLBiol,FLCatch
harvest(object, catch)

# S4 method for FLBiol
fbar(
  object,
  fisheries,
  range = unlist(dims(object)[c("min", "max")]),
  minfbar = range[1],
  maxfbar = range[2]
)

Arguments

object
Object containing population abundances, of class FLBiol.

catches
Object containing catches in number of the population represented by object, class FLFisheries.

...
Other things

fcb
A vector indicating the correspondance, by position of name, between object and the FLCatch elements inside catches, and of the same length.

units
Should output be in terms of fishing mortaloty ('f') or harvest rate ('hr').

catch
Object containing catches in number of the population represented by object, class FLFishery.

Value

An object of class FLQuant or FLQuants.

Details

The calculated fishing mortalities, or havest rates, are returned, in the case of harvest, disaggregated by fishery, as an FLQuants list. For a single FLFishery object, a single FLQuant is obtained.

See also

Author

The FLR Team

Examples

data(nsfishery)
harvests(ple, nsfleet)

#> $ple
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  1          2          3          4          5          6         
#>   1  9.8790e-03 4.3705e-03 4.3739e-03 5.2235e-03 6.1925e-03 7.1959e-03
#>   2  5.4901e-02 6.5060e-02 6.4679e-02 6.9595e-02 7.9832e-02 9.1391e-02
#>   3  2.0623e-01 1.6784e-01 1.5741e-01 1.6250e-01 1.7647e-01 1.9607e-01
#>   4  1.8865e-01 1.2258e-01 1.1488e-01 1.1491e-01 1.2391e-01 1.3488e-01
#>   5  1.5461e-01 7.4094e-02 6.4710e-02 6.1353e-02 6.0999e-02 6.3960e-02
#>   6  1.4561e-01 6.1755e-02 5.3233e-02 4.6770e-02 4.4989e-02 4.4656e-02
#>   7  1.4483e-01 6.1363e-02 5.2701e-02 4.6398e-02 4.2811e-02 4.2812e-02
#>   8  1.4481e-01 6.3506e-02 5.3286e-02 4.6832e-02 4.3478e-02 4.1932e-02
#>   9  1.3996e-05 6.5040e-02 5.5160e-02 4.7371e-02 4.3908e-02 4.2611e-02
#>   10 1.1952e-05 6.5572e-02 5.6765e-02 4.9971e-02 4.5947e-02 4.4241e-02
#>     year
#> age  7          8          9          10         11         12        
#>   1  8.1652e-03 9.2308e-03 1.0352e-02 1.1526e-02 1.2502e-02 1.3123e-02
#>   2  1.0445e-01 1.1927e-01 1.3553e-01 1.5238e-01 1.6843e-01 1.8088e-01
#>   3  2.1862e-01 2.4459e-01 2.7297e-01 3.0282e-01 3.3163e-01 3.5484e-01
#>   4  1.4901e-01 1.6465e-01 1.8160e-01 1.9936e-01 2.1670e-01 2.3074e-01
#>   5  6.6778e-02 7.0496e-02 7.4131e-02 7.8192e-02 8.2476e-02 8.6232e-02
#>   6  4.5530e-02 4.5644e-02 4.5956e-02 4.6430e-02 4.7359e-02 4.8427e-02
#>   7  4.2825e-02 4.3423e-02 4.3170e-02 4.3556e-02 4.4267e-02 4.5184e-02
#>   8  4.2461e-02 4.2466e-02 4.2951e-02 4.3086e-02 4.4018e-02 4.5011e-02
#>   9  4.1624e-02 4.2146e-02 4.2056e-02 4.2926e-02 4.3614e-02 4.4839e-02
#>   10 4.3347e-02 4.2453e-02 4.1926e-02 4.1942e-02 4.2686e-02 4.3674e-02
#>     year
#> age  13         14         15         16         17         18        
#>   1  1.3328e-02 1.3107e-02 1.2564e-02 1.1847e-02 1.1091e-02 1.0393e-02
#>   2  1.8738e-01 1.8732e-01 1.8176e-01 1.7283e-01 1.6270e-01 1.5297e-01
#>   3  3.6788e-01 3.6929e-01 3.6083e-01 3.4613e-01 3.2898e-01 3.1219e-01
#>   4  2.3895e-01 2.4019e-01 2.3544e-01 2.2693e-01 2.1689e-01 2.0697e-01
#>   5  8.8768e-02 8.9863e-02 8.9577e-02 8.8334e-02 8.6594e-02 8.4652e-02
#>   6  4.9515e-02 5.0529e-02 5.1434e-02 5.2143e-02 5.2604e-02 5.2778e-02
#>   7  4.6168e-02 4.7197e-02 4.8197e-02 4.9113e-02 4.9799e-02 5.0163e-02
#>   8  4.6075e-02 4.7127e-02 4.8151e-02 4.9070e-02 4.9804e-02 5.0207e-02
#>   9  4.5987e-02 4.7123e-02 4.8171e-02 4.9111e-02 4.9843e-02 5.0289e-02
#>   10 4.4837e-02 4.6043e-02 4.7214e-02 4.8271e-02 4.9122e-02 4.9671e-02
#>     year
#> age  19         20        
#>   1  9.8041e-03 9.3417e-03
#>   2  1.4455e-01 1.3782e-01
#>   3  2.9739e-01 2.8525e-01
#>   4  1.9811e-01 1.9071e-01
#>   5  8.2685e-02 8.0788e-02
#>   6  5.2630e-02 5.2175e-02
#>   7  5.0170e-02 4.9829e-02
#>   8  5.0225e-02 4.9883e-02
#>   9  5.0340e-02 5.0002e-02
#>   10 4.9859e-02 4.9681e-02
#> 
#> units:  NA 
#> 
#> $sol
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  1          2          3          4          5          6         
#>   1  9.8790e-03 1.1375e-02 1.0877e-02 1.1276e-02 1.2630e-02 1.4361e-02
#>   2  5.4901e-02 1.0399e-02 7.9192e-03 8.1282e-03 8.0660e-03 8.6881e-03
#>   3  2.0623e-01 9.8487e-03 8.5533e-03 6.7652e-03 6.9939e-03 6.6996e-03
#>   4  1.8865e-01 1.4459e-02 1.1195e-02 1.0345e-02 8.5354e-03 8.8427e-03
#>   5  1.5461e-01 8.2896e-03 8.4136e-03 6.5829e-03 6.0339e-03 4.8296e-03
#>   6  1.4561e-01 6.2099e-03 5.9212e-03 6.0463e-03 4.7904e-03 4.3704e-03
#>   7  1.4483e-01 5.3513e-03 5.2421e-03 5.1070e-03 5.4723e-03 4.5006e-03
#>   8  1.4481e-01 3.1702e-03 4.6196e-03 4.6327e-03 4.7589e-03 5.3278e-03
#>   9  1.3996e-05 1.6365e-03 2.7437e-03 4.0935e-03 4.3299e-03 4.6493e-03
#>   10 1.1952e-05 1.1040e-03 1.1393e-03 1.4942e-03 2.2909e-03 3.0193e-03
#>     year
#> age  7          8          9          10         11         12        
#>   1  1.6530e-02 1.9105e-02 2.2005e-02 2.5014e-02 2.8011e-02 3.0538e-02
#>   2  9.6692e-03 1.1138e-02 1.2861e-02 1.4758e-02 1.6557e-02 1.8270e-02
#>   3  7.0011e-03 7.5926e-03 8.5022e-03 9.5264e-03 1.0581e-02 1.1420e-02
#>   4  8.4251e-03 8.7298e-03 9.3398e-03 1.0296e-02 1.1315e-02 1.2233e-02
#>   5  4.7865e-03 4.3452e-03 4.2705e-03 4.3542e-03 4.5967e-03 4.8461e-03
#>   6  3.3893e-03 3.2120e-03 2.7678e-03 2.6002e-03 2.5506e-03 2.5984e-03
#>   7  4.1304e-03 3.1787e-03 2.9801e-03 2.5666e-03 2.4189e-03 2.3683e-03
#>   8  4.4344e-03 4.0661e-03 3.1190e-03 2.9479e-03 2.5685e-03 2.4334e-03
#>   9  5.2715e-03 4.3866e-03 4.0124e-03 3.1054e-03 2.9722e-03 2.6053e-03
#>   10 3.5490e-03 4.0794e-03 4.1427e-03 4.0893e-03 3.9009e-03 3.7705e-03
#>     year
#> age  13         14         15         16         17         18        
#>   1  3.2063e-02 3.2399e-02 3.1678e-02 3.0264e-02 2.8557e-02 2.6869e-02
#>   2  1.9606e-02 2.0257e-02 2.0233e-02 1.9684e-02 1.8843e-02 1.7919e-02
#>   3  1.2104e-02 1.2551e-02 1.2670e-02 1.2535e-02 1.2231e-02 1.1849e-02
#>   4  1.2794e-02 1.3158e-02 1.3328e-02 1.3273e-02 1.3081e-02 1.2809e-02
#>   5  5.0574e-03 5.1692e-03 5.2769e-03 5.3867e-03 5.4671e-03 5.5220e-03
#>   6  2.6700e-03 2.7582e-03 2.8360e-03 2.9501e-03 3.0895e-03 3.2190e-03
#>   7  2.4063e-03 2.4699e-03 2.5535e-03 2.6308e-03 2.7411e-03 2.8707e-03
#>   8  2.3879e-03 2.4278e-03 2.4905e-03 2.5702e-03 2.6386e-03 2.7347e-03
#>   9  2.4752e-03 2.4311e-03 2.4701e-03 2.5286e-03 2.5990e-03 2.6526e-03
#>   10 3.6254e-03 3.5112e-03 3.4271e-03 3.3680e-03 3.3196e-03 3.2703e-03
#>     year
#> age  19         20        
#>   1  2.5385e-02 2.4189e-02
#>   2  1.7054e-02 1.6327e-02
#>   3  1.1458e-02 1.1106e-02
#>   4  1.2510e-02 1.2221e-02
#>   5  5.5445e-03 5.5355e-03
#>   6  3.3260e-03 3.3976e-03
#>   7  2.9861e-03 3.0771e-03
#>   8  2.8456e-03 2.9407e-03
#>   9  2.7301e-03 2.8212e-03
#>   10 3.2107e-03 3.1417e-03
#> 
#> units:  NA 
#> 

harvest(ple, nsfleet[["bt"]], fcb="ple")

#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  1          2          3          4          5          6         
#>   1  1.9758e-02 1.5745e-02 1.5251e-02 1.6500e-02 1.8823e-02 2.1557e-02
#>   2  1.0980e-01 7.5459e-02 7.2598e-02 7.7723e-02 8.7898e-02 1.0008e-01
#>   3  4.1246e-01 1.7769e-01 1.6596e-01 1.6927e-01 1.8346e-01 2.0277e-01
#>   4  3.7729e-01 1.3704e-01 1.2608e-01 1.2526e-01 1.3245e-01 1.4373e-01
#>   5  3.0922e-01 8.2384e-02 7.3124e-02 6.7936e-02 6.7033e-02 6.8790e-02
#>   6  2.9123e-01 6.7965e-02 5.9155e-02 5.2816e-02 4.9780e-02 4.9026e-02
#>   7  2.8967e-01 6.6715e-02 5.7943e-02 5.1505e-02 4.8283e-02 4.7312e-02
#>   8  2.8962e-01 6.6676e-02 5.7906e-02 5.1465e-02 4.8237e-02 4.7260e-02
#>   9  1.3996e-05 6.6676e-02 5.7904e-02 5.1465e-02 4.8238e-02 4.7260e-02
#>   10 1.1952e-05 6.6676e-02 5.7904e-02 5.1465e-02 4.8238e-02 4.7260e-02
#>     year
#> age  7          8          9          10         11         12        
#>   1  2.4696e-02 2.8336e-02 3.2357e-02 3.6540e-02 4.0513e-02 4.3660e-02
#>   2  1.1412e-01 1.3041e-01 1.4840e-01 1.6714e-01 1.8499e-01 1.9915e-01
#>   3  2.2562e-01 2.5219e-01 2.8148e-01 3.1235e-01 3.4222e-01 3.6626e-01
#>   4  1.5743e-01 1.7338e-01 1.9094e-01 2.0965e-01 2.2801e-01 2.4297e-01
#>   5  7.1564e-02 7.4841e-02 7.8401e-02 8.2547e-02 8.7073e-02 9.1078e-02
#>   6  4.8920e-02 4.8856e-02 4.8723e-02 4.9030e-02 4.9909e-02 5.1025e-02
#>   7  4.6956e-02 4.6602e-02 4.6150e-02 4.6123e-02 4.6686e-02 4.7552e-02
#>   8  4.6895e-02 4.6533e-02 4.6070e-02 4.6033e-02 4.6587e-02 4.7445e-02
#>   9  4.6896e-02 4.6533e-02 4.6069e-02 4.6032e-02 4.6587e-02 4.7445e-02
#>   10 4.6896e-02 4.6533e-02 4.6069e-02 4.6032e-02 4.6587e-02 4.7445e-02
#>     year
#> age  13         14         15         16         17         18        
#>   1  4.5391e-02 4.5506e-02 4.4242e-02 4.2110e-02 3.9648e-02 3.7262e-02
#>   2  2.0699e-01 2.0758e-01 2.0199e-01 1.9251e-01 1.8154e-01 1.7088e-01
#>   3  3.7999e-01 3.8184e-01 3.7350e-01 3.5867e-01 3.4121e-01 3.2404e-01
#>   4  2.5174e-01 2.5335e-01 2.4876e-01 2.4021e-01 2.2997e-01 2.1978e-01
#>   5  9.3826e-02 9.5032e-02 9.4854e-02 9.3721e-02 9.2061e-02 9.0174e-02
#>   6  5.2185e-02 5.3287e-02 5.4270e-02 5.5093e-02 5.5694e-02 5.5997e-02
#>   7  4.8574e-02 4.9667e-02 5.0750e-02 5.1743e-02 5.2540e-02 5.3033e-02
#>   8  4.8463e-02 4.9555e-02 5.0642e-02 5.1640e-02 5.2443e-02 5.2942e-02
#>   9  4.8462e-02 4.9555e-02 5.0641e-02 5.1639e-02 5.2442e-02 5.2941e-02
#>   10 4.8462e-02 4.9555e-02 5.0641e-02 5.1639e-02 5.2442e-02 5.2941e-02
#>     year
#> age  19         20        
#>   1  3.5190e-02 2.8723e-05
#>   2  1.6160e-01 3.7250e-03
#>   3  3.0884e-01 4.3259e-02
#>   4  2.1062e-01 4.9026e-02
#>   5  8.8229e-02 5.1948e-02
#>   6  5.5956e-02 5.2750e-02
#>   7  5.3157e-02 5.2821e-02
#>   8  5.3070e-02 5.2823e-02
#>   9  5.3070e-02 5.2823e-02
#>   10 5.3070e-02 5.2823e-02
#> 
#> units:  f 

harvest(ple, nsfleet)

#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  1          2          3          4          5          6         
#>   1  1.9758e-02 1.5745e-02 1.5251e-02 1.6500e-02 1.8823e-02 2.1557e-02
#>   2  1.0980e-01 7.5459e-02 7.2598e-02 7.7723e-02 8.7898e-02 1.0008e-01
#>   3  4.1246e-01 1.7769e-01 1.6596e-01 1.6927e-01 1.8346e-01 2.0277e-01
#>   4  3.7729e-01 1.3704e-01 1.2608e-01 1.2526e-01 1.3245e-01 1.4373e-01
#>   5  3.0922e-01 8.2384e-02 7.3124e-02 6.7936e-02 6.7033e-02 6.8790e-02
#>   6  2.9123e-01 6.7965e-02 5.9155e-02 5.2816e-02 4.9780e-02 4.9026e-02
#>   7  2.8967e-01 6.6715e-02 5.7943e-02 5.1505e-02 4.8283e-02 4.7312e-02
#>   8  2.8962e-01 6.6676e-02 5.7906e-02 5.1465e-02 4.8237e-02 4.7260e-02
#>   9  2.7993e-05 6.6676e-02 5.7904e-02 5.1465e-02 4.8238e-02 4.7260e-02
#>   10 2.3904e-05 6.6676e-02 5.7904e-02 5.1465e-02 4.8238e-02 4.7260e-02
#>     year
#> age  7          8          9          10         11         12        
#>   1  2.4696e-02 2.8336e-02 3.2357e-02 3.6540e-02 4.0513e-02 4.3660e-02
#>   2  1.1412e-01 1.3041e-01 1.4840e-01 1.6714e-01 1.8499e-01 1.9915e-01
#>   3  2.2562e-01 2.5219e-01 2.8148e-01 3.1235e-01 3.4222e-01 3.6626e-01
#>   4  1.5743e-01 1.7338e-01 1.9094e-01 2.0965e-01 2.2801e-01 2.4297e-01
#>   5  7.1564e-02 7.4841e-02 7.8401e-02 8.2547e-02 8.7073e-02 9.1078e-02
#>   6  4.8920e-02 4.8856e-02 4.8723e-02 4.9030e-02 4.9909e-02 5.1025e-02
#>   7  4.6956e-02 4.6602e-02 4.6150e-02 4.6123e-02 4.6686e-02 4.7552e-02
#>   8  4.6895e-02 4.6533e-02 4.6070e-02 4.6033e-02 4.6587e-02 4.7445e-02
#>   9  4.6896e-02 4.6533e-02 4.6069e-02 4.6032e-02 4.6587e-02 4.7445e-02
#>   10 4.6896e-02 4.6533e-02 4.6069e-02 4.6032e-02 4.6587e-02 4.7445e-02
#>     year
#> age  13         14         15         16         17         18        
#>   1  4.5391e-02 4.5506e-02 4.4242e-02 4.2110e-02 3.9648e-02 3.7262e-02
#>   2  2.0699e-01 2.0758e-01 2.0199e-01 1.9251e-01 1.8154e-01 1.7088e-01
#>   3  3.7999e-01 3.8184e-01 3.7350e-01 3.5867e-01 3.4121e-01 3.2404e-01
#>   4  2.5174e-01 2.5335e-01 2.4876e-01 2.4021e-01 2.2997e-01 2.1978e-01
#>   5  9.3826e-02 9.5032e-02 9.4854e-02 9.3721e-02 9.2061e-02 9.0174e-02
#>   6  5.2185e-02 5.3287e-02 5.4270e-02 5.5093e-02 5.5694e-02 5.5997e-02
#>   7  4.8574e-02 4.9667e-02 5.0750e-02 5.1743e-02 5.2540e-02 5.3033e-02
#>   8  4.8463e-02 4.9555e-02 5.0642e-02 5.1640e-02 5.2443e-02 5.2942e-02
#>   9  4.8462e-02 4.9555e-02 5.0641e-02 5.1639e-02 5.2442e-02 5.2941e-02
#>   10 4.8462e-02 4.9555e-02 5.0641e-02 5.1639e-02 5.2442e-02 5.2941e-02
#>     year
#> age  19         20        
#>   1  3.5190e-02 3.3530e-02
#>   2  1.6160e-01 1.5415e-01
#>   3  3.0884e-01 2.9636e-01
#>   4  2.1062e-01 2.0293e-01
#>   5  8.8229e-02 8.6324e-02
#>   6  5.5956e-02 5.5572e-02
#>   7  5.3157e-02 5.2906e-02
#>   8  5.3070e-02 5.2824e-02
#>   9  5.3070e-02 5.2823e-02
#>   10 5.3070e-02 5.2823e-02
#> 
#> units:  f 

harvest(ple, nsfleet[["bt"]][["ple"]])

#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  1          2          3          4          5          6         
#>   1  1.9758e-02 1.5745e-02 1.5251e-02 1.6500e-02 1.8823e-02 2.1557e-02
#>   2  1.0980e-01 7.5459e-02 7.2598e-02 7.7723e-02 8.7898e-02 1.0008e-01
#>   3  4.1246e-01 1.7769e-01 1.6596e-01 1.6927e-01 1.8346e-01 2.0277e-01
#>   4  3.7729e-01 1.3704e-01 1.2608e-01 1.2526e-01 1.3245e-01 1.4373e-01
#>   5  3.0922e-01 8.2384e-02 7.3124e-02 6.7936e-02 6.7033e-02 6.8790e-02
#>   6  2.9123e-01 6.7965e-02 5.9155e-02 5.2816e-02 4.9780e-02 4.9026e-02
#>   7  2.8967e-01 6.6715e-02 5.7943e-02 5.1505e-02 4.8283e-02 4.7312e-02
#>   8  2.8962e-01 6.6676e-02 5.7906e-02 5.1465e-02 4.8237e-02 4.7260e-02
#>   9  1.3996e-05 6.6676e-02 5.7904e-02 5.1465e-02 4.8238e-02 4.7260e-02
#>   10 1.1952e-05 6.6676e-02 5.7904e-02 5.1465e-02 4.8238e-02 4.7260e-02
#>     year
#> age  7          8          9          10         11         12        
#>   1  2.4696e-02 2.8336e-02 3.2357e-02 3.6540e-02 4.0513e-02 4.3660e-02
#>   2  1.1412e-01 1.3041e-01 1.4840e-01 1.6714e-01 1.8499e-01 1.9915e-01
#>   3  2.2562e-01 2.5219e-01 2.8148e-01 3.1235e-01 3.4222e-01 3.6626e-01
#>   4  1.5743e-01 1.7338e-01 1.9094e-01 2.0965e-01 2.2801e-01 2.4297e-01
#>   5  7.1564e-02 7.4841e-02 7.8401e-02 8.2547e-02 8.7073e-02 9.1078e-02
#>   6  4.8920e-02 4.8856e-02 4.8723e-02 4.9030e-02 4.9909e-02 5.1025e-02
#>   7  4.6956e-02 4.6602e-02 4.6150e-02 4.6123e-02 4.6686e-02 4.7552e-02
#>   8  4.6895e-02 4.6533e-02 4.6070e-02 4.6033e-02 4.6587e-02 4.7445e-02
#>   9  4.6896e-02 4.6533e-02 4.6069e-02 4.6032e-02 4.6587e-02 4.7445e-02
#>   10 4.6896e-02 4.6533e-02 4.6069e-02 4.6032e-02 4.6587e-02 4.7445e-02
#>     year
#> age  13         14         15         16         17         18        
#>   1  4.5391e-02 4.5506e-02 4.4242e-02 4.2110e-02 3.9648e-02 3.7262e-02
#>   2  2.0699e-01 2.0758e-01 2.0199e-01 1.9251e-01 1.8154e-01 1.7088e-01
#>   3  3.7999e-01 3.8184e-01 3.7350e-01 3.5867e-01 3.4121e-01 3.2404e-01
#>   4  2.5174e-01 2.5335e-01 2.4876e-01 2.4021e-01 2.2997e-01 2.1978e-01
#>   5  9.3826e-02 9.5032e-02 9.4854e-02 9.3721e-02 9.2061e-02 9.0174e-02
#>   6  5.2185e-02 5.3287e-02 5.4270e-02 5.5093e-02 5.5694e-02 5.5997e-02
#>   7  4.8574e-02 4.9667e-02 5.0750e-02 5.1743e-02 5.2540e-02 5.3033e-02
#>   8  4.8463e-02 4.9555e-02 5.0642e-02 5.1640e-02 5.2443e-02 5.2942e-02
#>   9  4.8462e-02 4.9555e-02 5.0641e-02 5.1639e-02 5.2442e-02 5.2941e-02
#>   10 4.8462e-02 4.9555e-02 5.0641e-02 5.1639e-02 5.2442e-02 5.2941e-02
#>     year
#> age  19         20        
#>   1  3.5190e-02 2.8723e-05
#>   2  1.6160e-01 3.7250e-03
#>   3  3.0884e-01 4.3259e-02
#>   4  2.1062e-01 4.9026e-02
#>   5  8.8229e-02 5.1948e-02
#>   6  5.5956e-02 5.2750e-02
#>   7  5.3157e-02 5.2821e-02
#>   8  5.3070e-02 5.2823e-02
#>   9  5.3070e-02 5.2823e-02
#>   10 5.3070e-02 5.2823e-02
#> 
#> units:  f 

fbar(ple, fisheries=nsfleet, minfbar=3, maxfbar=6)

#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> age   1       2       3       4       5       6       7       8       9      
#>   all 0.34755 0.11627 0.10608 0.10382 0.10818 0.11608 0.12588 0.13732 0.14989
#>      year
#> age   10      11      12      13      14      15      16      17      18     
#>   all 0.16339 0.17680 0.18783 0.19444 0.19588 0.19285 0.18692 0.17974 0.17250
#>      year
#> age   19      20     
#>   all 0.16591 0.16030
#> 
#> units:  f 

fbar(ple, fisheries=nsfleet[["bt"]], minfbar=3, maxfbar=6)

#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> age   1        2        3        4        5        6        7        8       
#>   all 0.347550 0.116269 0.106080 0.103819 0.108181 0.116077 0.125885 0.137316
#>      year
#> age   9        10       11       12       13       14       15       16      
#>   all 0.149886 0.163394 0.176802 0.187834 0.194435 0.195875 0.192846 0.186922
#>      year
#> age   17       18       19       20      
#>   all 0.179736 0.172499 0.165912 0.049246
#> 
#> units:  f