FLCatch constructor

Make an FLCatch object.

Catch data for a single species or stock unit is handled by the FLCatch class. Data is separated as landings and discards by age, in numbers, with the corresponding mean weights at age.

FLCatch(object, ...)

catch.sel(object, ...) <- value

discards.ratio(object, ...)

lrevenue(object, ...)

# S4 method for FLCatch
landings.n(object)

# S4 method for FLCatch,numeric
landings.n(object) <- value

# S4 method for FLCatch,numeric
landings.wt(object) <- value

# S4 method for FLCatch,numeric
discards.n(object) <- value

# S4 method for FLCatch,numeric
discards.wt(object) <- value

# S4 method for FLCatch,numeric
catch.sel(object) <- value

# S4 method for FLCatch,numeric
price(object) <- value

# S4 method for FLCatch,numeric
catch.q(object) <- value

# S4 method for FLQuant
FLCatch(object, ...)

# S4 method for missing
FLCatch(object, ...)

# S4 method for FLCatch
lrevenue(object)

# S4 method for FLCatch,missing
plot(x, y, ...)

Arguments

object

Either an FLQuant (to determine the size of the FLQuant slots) or missing

...

Other things

value

Replacement value

x

whatever

y

whatever

Value

An FLCatch object

Details

Make an FLCatch object.

This is class is used inside FLFishery to store the catches of a single stock or species caught by that fleet.

Slots

catch.q

Parameters of the catchability function, (FLPar).

catch.sel

Selectivity at age as proportions over fully-selected ages, (FLQuant).

desc

Description of the data contents and origin, (character).

discards.n

Discards at age in numbers, (FLQuant).

discards.wt

Mean weight-at-age in the discards, (FLQuant).

name

Name of the object, e.g. species or stock code, (character).

landings.n

Landings at age in numbers, (FLQuant).

landings.wt

Mean weight-at-age in the landings, (FLQuant).

price

Mean price by age per unit of weight, (FLQuant).

range

Ranges of age and years, plusgroup, (numeric).

Validity

Length of dimensions 2:5

All FLQuant slots must share dimensions 2 to 5.

iter dim of length 1 or N

The 6th dimension in all FLQuant and FLPar slots must be 1 or N, where N is the same value for the whole object.

Length of dimensions 2:5

All FLQuant slots must share dimensions 2 to 5.

You can inspect the class validity function by using getValidity(getClassDef('FLCatch'))

Accessors

All slots in the class have accessor and replacement methods defined that allow retrieving and substituting individual slots.

The values passed for replacement need to be of the class of that slot. A numeric vector can also be used when replacing FLQuant slots, and the vector will be used to substitute the values in the slot, but not its other attributes.

Constructor

A construction method exists for this class that can take named arguments for any of its slots. All slots are then created to match the requirements of the class validity. If an unnamed FLQuant object is provided, this is used for sizing but not stored in any slot.

Methods

Methods exist for various calculations based on values stored in the class:

landings

Total landings as sum on 'age' of landings.n times landings.wt.

discards

Total discards as sum on 'age' of discards.n times discards.wt.

landings.sel

Selectivity at age in the landings as proportions over fully-selected ages, (FLQuant).

discards.sel

Selectivity at age in the discards as proportions over fully-selected ages, (FLQuant).

catch.n

Catch at age in numbers as landings.n plus discards.n.

catch.wt

Weighted average of landings.wt and discards.wt.

catch

Total catch as sum of landings and discards.

discards.ratio

Proportion at age of discards in catch.

plot

Standard plot for the FLCatch class.

See also

FLComp

FLCatches, FLFishery

Examples

data(ple4)

# EXTRACT data from FLCore ple4, fake prices
fca <- FLCatch(name='PLE', desc='All NS PLE catches',
  landings.n=landings.n(ple4), landings.wt=landings.wt(ple4),
  discards.n=discards.n(ple4), discards.wt=discards.wt(ple4),
  price=landings.wt(ple4) * 23, catch.q=FLPar(q=1),
  catch.sel=catch.sel(ple4))

# Calculations
landings(fca)

#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> age   1957   1958   1959   1960   1961   1962   1963   1964   1965   1966  
#>   all  70563  73354  79300  87541  85984  87472 107118 110540  97143 101834
#>      year
#> age   1967   1968   1969   1970   1971   1972   1973   1974   1975   1976  
#>   all 108819 111534 121651 130342 113944 122843 130429 112540 108536 113670
#>      year
#> age   1977   1978   1979   1980   1981   1982   1983   1984   1985   1986  
#>   all 119188 113984 145347 139951 139747 154547 144038 156147 159838 165347
#>      year
#> age   1987   1988   1989   1990   1991   1992   1993   1994   1995   1996  
#>   all 153670 154475 169818 156240 148004 125190 117113 110392  98356  81673
#>      year
#> age   1997   1998   1999   2000   2001   2002   2003   2004   2005   2006  
#>   all  83048  71534  80662  81148  81963  70217  66502  61436  55700  57943
#>      year
#> age   2007   2008  
#>   all  49744  48874
#> 
#> units:  t 

catch.n(fca)

#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  1957    1958    1959    1960    1961    1962    1963    1964    1965   
#>   1    32356   66199  116086   73939   75578   51265   90913   66035   43708
#>   2    49911   80681  144327  173852  146873  183468  140523  167982  435879
#>   3    69038   45860   76829  106824  122406  123475  151249  104642  101464
#>   4    45627   64619   36896   52019   68444   88495   86890   98560   56592
#>   5    32732   36249   45836   22388   34311   50859   56476   60419   44597
#>   6     8910   19659   23042   27848   12889   20814   31987   37414   30418
#>   7    11029    8178   13873   14186   14680    9060   12848   15819   18361
#>   8     9028    8000    6408    9013    9748    9035    6833    6595    8554
#>   9     4973    6110    6596    5087    5996    5257    7047    3980    4213
#>   10   10859   13148   16180   15153   14660   12801   16592   16886   17587
#>     year
#> age  1966    1967    1968    1969    1970    1971    1972    1973    1974   
#>   1    38496   20199   73971   85195  123645   69356   72235  133620  213362
#>   2   167269  139527   81666   82395  169774  126559   91998   81548  331531
#>   3   414367  117836   87037   61934   78921   90636   96098   63107   59200
#>   4    65887  265413   64670   55418   57440   36150   58620   73738   42410
#>   5    38069   42006  146055   36358   45747   26485   23610   43544   41686
#>   6    27495   22313   17197  104038   35235   22984   17071   13884   19775
#>   7    16500   16175   10341   10410   78886   16913   14380    8885    8005
#>   8    10784    8004   10102    6086    6311   29730   10903    9848    6321
#>   9     6467    6728    3925    8192    4185    6414   18585    6084    5568
#>   10   14928   11175   13365   16092   14840   16910   15651   23978   21980
#>     year
#> age  1975    1976    1977    1978    1979    1980    1981    1982    1983   
#>   1   245950  186699  259848  228015  294484  227350  134395  414641  262614
#>   2   308254  174564  160665  214691  273115  187465  294168  252348  556026
#>   3   252159  148703  122606   89600  176441  134673  124146  213631  145750
#>   4    36069  114411   99565   65116   67344   78210   57977   70653  101311
#>   5    25672   13953   93022   51346   48475   25167   36176   28871   30163
#>   6    21311    9945    9271   36080   40242   18068   12469   16824   12978
#>   7    11873    9120    5912    5940   24228   13761    9564    7589    8216
#>   8     5923    6391    5022    3352    4161    8458    8092    5470    4193
#>   9     4106    2947    4061    2419    2807    1864    4874    4482    3013
#>   10   19695   12552    9191    7468    9288    5377    5903    8653    8287
#>     year
#> age  1984    1985    1986    1987    1988    1989    1990    1991    1992   
#>   1   310783  405506 1119019  361519  348597  214552  146864  184587  142165
#>   2   376742  302760  558090 1459325  623255  532602  311831  344841  263573
#>   3   326860  179882  212227  296920  710060  299105  266440  225170  179701
#>   4    78078  187424  120271  112666  135502  317172  139099  156894  106842
#>   5    38960   32720   85930   66106   48262   60133  164825   78489   77010
#>   6    13730   15622   24195   34976   25268   19362   26934   78496   34286
#>   7     6465    6871    9299   11452   16774   10567    8129   14910   29547
#>   8     5544    3650    4490    4341    5381    7561    4216    5233    6970
#>   9     2720    2698    2733    2154    3162    2120    3451    3141    3200
#>   10    6565    5798    6950    5478    6233    5580    3808    5591    6928
#>     year
#> age  1993    1994    1995    1996    1997    1998    1999    2000    2001   
#>   1    99832   63516  126614  112354  129545  104734  127870  106056   34804
#>   2   208032  140851  119251  373561  553773  676651  217090  187647  388808
#>   3   146463  137320   97151   91988  280776  260551  387740   91904  240444
#>   4    84252   81274   79230   50289   44257  109683   94726  230201  123671
#>   5    53040   39364   37162   32581   22812   17010   24390   16625  144551
#>   6    29217   20532   18073   14576   13626    6465    6887    9655    7048
#>   7    13472   15323    9772   10605    6362    4986    2783    2399    4464
#>   8    11272    6399    4366    4528    3632    2506    2246    1198    1128
#>   9     3645    5368    2336    2624    2179    1761    1521     940     637
#>   10    5883    5433    3753    4892    4181    3119    3093    2097    2309
#>     year
#> age  2002    2003    2004    2005    2006    2007    2008   
#>   1   311003   68309  233397   96478  220828   78598  135786
#>   2   193978  560788  191141  342682  244064  224345  263599
#>   3   136274  101562  221343   72076  177680  106621   80258
#>   4    61739   68836   43238  103120   35910   57788   44207
#>   5    38681   31339   22201   21746   45525   16408   31758
#>   6    38638   16454   11650   15610    6965   24293    6217
#>   7     4763   22821    4830    5489    5721    2663   15634
#>   8     1789    1609    4556    2615    2247    3505    1803
#>   9      742     650     412    2223    1235     543     673
#>   10    1586     859     540     613    1319    1259     896
#> 
#> units:  1000 

catch.wt(fca)

#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  1957     1958     1959     1960     1961     1962     1963     1964    
#>   1  0.046000 0.049000 0.053000 0.047000 0.046000 0.044000 0.049000 0.034000
#>   2  0.109023 0.105210 0.119391 0.113130 0.099479 0.098251 0.107783 0.124247
#>   3  0.213031 0.195285 0.193155 0.204511 0.180903 0.179610 0.175238 0.204511
#>   4  0.284385 0.271854 0.264005 0.288912 0.305796 0.265923 0.309197 0.272009
#>   5  0.386484 0.348902 0.351582 0.379534 0.408241 0.383857 0.399080 0.381492
#>   6  0.505611 0.480920 0.482110 0.483042 0.513842 0.520182 0.541447 0.488312
#>   7  0.591892 0.545508 0.604854 0.604999 0.613395 0.550927 0.636038 0.632705
#>   8  0.654079 0.653972 0.636527 0.687545 0.681438 0.668845 0.679868 0.705244
#>   9  0.439755 0.707140 0.766285 0.728818 0.824631 0.751001 0.728794 0.742521
#>   10 1.108295 1.054905 1.021313 1.101008 1.087871 1.089642 1.047979 1.012463
#>     year
#> age  1965     1966     1967     1968     1969     1970     1971     1972    
#>   1  0.040000 0.040000 0.038000 0.062000 0.055006 0.051157 0.059056 0.068872
#>   2  0.074899 0.103375 0.109685 0.115856 0.174456 0.129876 0.160326 0.208282
#>   3  0.214331 0.149364 0.189797 0.226458 0.282870 0.263496 0.280962 0.295307
#>   4  0.314895 0.318863 0.236808 0.278844 0.293982 0.343297 0.400173 0.417686
#>   5  0.383521 0.435071 0.430126 0.347836 0.376170 0.384340 0.458767 0.499705
#>   6  0.471450 0.491629 0.553895 0.531091 0.431576 0.429678 0.529129 0.554980
#>   7  0.542063 0.569186 0.608835 0.607190 0.606340 0.486216 0.559926 0.625212
#>   8  0.666768 0.635371 0.675124 0.613365 0.693111 0.654783 0.627076 0.664288
#>   9  0.730126 0.702573 0.752630 0.705987 0.696285 0.724939 0.722119 0.693080
#>   10 0.891599 0.950224 0.997522 0.937469 0.945344 0.868761 0.920023 0.965363
#>     year
#> age  1973     1974     1975     1976     1977     1978     1979     1980    
#>   1  0.047285 0.059485 0.070776 0.090775 0.073273 0.071824 0.069741 0.057793
#>   2  0.207399 0.120167 0.150853 0.180195 0.216046 0.186920 0.188034 0.196467
#>   3  0.349811 0.355104 0.208088 0.264594 0.244518 0.307195 0.294753 0.347567
#>   4  0.423437 0.418549 0.413929 0.354364 0.317391 0.353056 0.336748 0.404721
#>   5  0.501593 0.489456 0.522874 0.522191 0.396373 0.417079 0.425723 0.477425
#>   6  0.564744 0.572571 0.621438 0.607570 0.552086 0.469304 0.471324 0.549762
#>   7  0.635713 0.631496 0.675687 0.657268 0.648046 0.587232 0.549059 0.595974
#>   8  0.658829 0.718599 0.746868 0.722892 0.722165 0.662444 0.673845 0.670978
#>   9  0.711368 0.733116 0.831860 0.759806 0.716156 0.748263 0.794638 0.782469
#>   10 0.884395 0.959747 1.082312 1.005046 0.979675 0.915809 0.958620 1.026708
#>     year
#> age  1981     1982     1983     1984     1985     1986     1987     1988    
#>   1  0.050360 0.057807 0.054669 0.055061 0.056056 0.051250 0.044000 0.044000
#>   2  0.183023 0.151248 0.150788 0.147479 0.167321 0.140162 0.112277 0.101697
#>   3  0.332426 0.310109 0.273235 0.261412 0.262784 0.272772 0.216260 0.196281
#>   4  0.421914 0.423076 0.376695 0.319147 0.328625 0.309557 0.344810 0.273222
#>   5  0.510332 0.515070 0.503613 0.472596 0.451360 0.416246 0.393201 0.441804
#>   6  0.566345 0.609581 0.598260 0.600393 0.563800 0.480560 0.495832 0.502289
#>   7  0.614579 0.668149 0.672781 0.673317 0.664375 0.666696 0.575739 0.598740
#>   8  0.653244 0.716449 0.766195 0.713676 0.714350 0.741761 0.718934 0.688355
#>   9  0.737698 0.742611 0.809921 0.823927 0.787842 0.842506 0.818676 0.800618
#>   10 1.024814 0.989912 0.977658 1.018571 1.000805 1.000688 0.977816 0.998985
#>     year
#> age  1989     1990     1991     1992     1993     1994     1995     1996    
#>   1  0.048091 0.056246 0.059272 0.059592 0.065386 0.065533 0.072869 0.054801
#>   2  0.117269 0.130487 0.148260 0.146382 0.158028 0.177893 0.181464 0.139483
#>   3  0.213113 0.208319 0.207016 0.222763 0.245734 0.251585 0.280543 0.264973
#>   4  0.287375 0.286435 0.266293 0.271638 0.301019 0.328338 0.334307 0.338074
#>   5  0.362951 0.355638 0.340895 0.327257 0.343305 0.382913 0.396031 0.410843
#>   6  0.523569 0.438740 0.435741 0.411947 0.411796 0.435768 0.449573 0.476535
#>   7  0.594220 0.587803 0.508682 0.521339 0.505929 0.488823 0.524551 0.491304
#>   8  0.660244 0.681340 0.646450 0.593992 0.616296 0.594962 0.607009 0.579916
#>   9  0.780468 0.749277 0.719670 0.701991 0.704199 0.712785 0.729205 0.708895
#>   10 0.928990 0.988549 0.886882 0.874771 0.835903 0.882795 0.902195 0.844397
#>     year
#> age  1997     1998     1999     2000     2001     2002     2003     2004    
#>   1  0.043132 0.049188 0.055801 0.064984 0.091461 0.058690 0.069618 0.057360
#>   2  0.129672 0.093633 0.102143 0.122586 0.135140 0.130217 0.124168 0.123887
#>   3  0.206602 0.204263 0.197135 0.205278 0.193655 0.218882 0.224444 0.219356
#>   4  0.358965 0.294967 0.259078 0.267185 0.227299 0.283772 0.280163 0.300012
#>   5  0.450560 0.484290 0.445941 0.401026 0.298546 0.332798 0.341227 0.379188
#>   6  0.517609 0.594337 0.536878 0.469879 0.469842 0.431328 0.400024 0.425953
#>   7  0.598161 0.622975 0.621207 0.611979 0.580124 0.487882 0.412311 0.502313
#>   8  0.611056 0.683993 0.672481 0.592818 0.700604 0.677245 0.640330 0.551459
#>   9  0.677518 0.688914 0.741504 0.729327 0.787163 0.745190 0.750456 0.789245
#>   10 0.916741 0.899651 0.801638 0.861727 0.792664 0.881252 0.837241 0.860880
#>     year
#> age  2005     2006     2007     2008    
#>   1  0.069222 0.062313 0.060347 0.058530
#>   2  0.116426 0.139194 0.111512 0.122204
#>   3  0.210045 0.210275 0.221105 0.240195
#>   4  0.296996 0.294857 0.314374 0.320583
#>   5  0.349228 0.385415 0.361766 0.387632
#>   6  0.340212 0.396499 0.377003 0.434290
#>   7  0.456296 0.439897 0.514525 0.357386
#>   8  0.544459 0.464998 0.347238 0.460998
#>   9  0.603346 0.532624 0.590289 0.640373
#>   10 0.887783 0.755080 0.618717 0.636754
#> 
#> units:  kg