Classes and Methods for Modeling of Fishing Fleets

FLFishery defines a set of classes to load, manipulate and combine landings discards data from fishing fleets catching individuals of one or more stocks.

Fishing fleets consisting of a number of boats operating homogeneously can be modelled using the FLFishery class. All boats in the fleet must have a common gear configuration during each time step and area (no metiers).

FLFishery(object, ...)

capacity(object, ...)

capacity(object, ...) <- value

crewshare(object, ...)

crewshare(object, ...) <- value

hperiod(object, ...)

hperiod(object, ...) <- value

orevenue(object, ...)

orevenue(object, ...) <- value

# S4 method for FLFishery
capacity(object)

# S4 method for FLFishery,FLQuant
capacity(object) <- value

# S4 method for FLFishery,numeric
capacity(object) <- value

# S4 method for FLFishery
hperiod(object)

# S4 method for FLFishery,FLQuant
hperiod(object) <- value

# S4 method for FLFishery,numeric
hperiod(object) <- value

# S4 method for FLFishery,ANY
effort(object, compute = TRUE)

# S4 method for FLFishery,FLQuant
effort(object) <- value

# S4 method for FLFishery
vcost(object, compute = TRUE)

# S4 method for FLFishery,FLQuant
vcost(object) <- value

# S4 method for FLFishery
fcost(object, compute = TRUE)

# S4 method for FLFishery,FLQuant
fcost(object) <- value

# S4 method for FLFishery,numeric
fcost(object) <- value

# S4 method for FLFishery
orevenue(object)

# S4 method for FLFishery,FLQuant
orevenue(object) <- value

# S4 method for FLFishery,numeric
orevenue(object) <- value

# S4 method for FLFishery
crewshare(object, compute = TRUE)

# S4 method for FLFishery,predictModel
crewshare(object) <- value

# S4 method for FLFishery
landings(object, catch = seq(object))

# S4 method for FLFishery
discards(object, catch = seq(object))

# S4 method for FLFishery
catch(object, catch = seq(object))

# S4 method for FLFishery
catch.n(object, pos = names(object))

# S4 method for FLFishery
landings.wt(object, pos = names(object))

# S4 method for FLFishery
catch.wt(object, pos = names(object))

# S4 method for FLFishery
catch.wt(object, pos = names(object))

# S4 method for list
FLFishery(object, ...)

# S4 method for FLCatch
FLFishery(object, ...)

# S4 method for missing
FLFishery(object, ...)

# S4 method for FLFishery,ANY,missing,ANY
[(x, i)

# S4 method for FLFishery,numeric,missing,FLCatch
[[(x, i) <- value

# S4 method for FLFishery,character,missing,FLCatch
[[(x, i) <- value

# S4 method for FLFishery
summary(object)

# S4 method for FLFishery
lrevenue(object)

# S4 method for FLFishery
revenue(object)

# S4 method for FLFishery
cost(object)

# S4 method for FLFishery
profit(object)

# S4 method for FLFishery
ccost(object)

# S4 method for FLFishery
propagate(object, iter, fill.iter = TRUE)

# S4 method for FLFishery
iter(obj, iter)

# S4 method for FLFishery,FLFishery
iter(object, iter) <- value

Arguments

object

Object on which to assign value

...

Other things

value

Object to assign

compute

Carry out formula calculation (TRUE) or return full slot (FALSE).

x

Object to be subset

i

Element to be extracted, by name (character) or position (numeric).

iter

Position (numeric) or name (character) of the iter(s) to be extracted (iter), or number of iters to be created (propagate).

fill.iter

Should the object content be copied across the new iters, logical.

obj

Object on which to apply method

Details

Fisheries are defined here as a group of vessels catching from a number of stocks

What do you say what I can ever do for you What are we gonna do to pass the time What do you care when you find that life's unfair Equality is just a state of mind Believe whatever is right what's right for you tonight You know where to draw the line

Slots

FLFishery objects inherit from FLCatches with a number of slots added.

.Data

The list of FLCatch object with catch data per stock, (FLCatches).

name

Name of the object, e.g. species or stock code, (character).

desc

Description of the data contents and origin, (character).

range

Ranges of age and years, plusgroup, (numeric).

capacity

Number of boats in the fleet, (FLQuant).

effort

Mean effort per boat applied by the fleet, (FLQuant).

hperiod

Start and end of fishing within each time step, as proportions. An FLQuant object with dimnames `quant=c('start', 'end')` in the first dimension.

vcost

Variable costs per unit of effort, (FLQuant).

fcost

Variable costs per unit of effort, (FLQuant).

orevenue

Revenues obtained from sources other than landings, (FLQuant).

crewshare

Formula, parameter values and inputs to calculate the crew costs, (predictModel).

Validity

You can inspect the class validity function by using getValidity(getClassDef('FLFishery'))

Accessors

All slots in the class have accessor and replacement methods defined that allow retrieving and substituting individual slots.

The values passed for replacement need to be of the class of that slot. A numeric vector can also be used when replacing FLQuant slots, and the vector will be used to substitute the values in the slot, but not its other attributes.

Constructor

A construction method exists for this class that can take named arguments for any of its slots. All slots are then created to match the requirements of the class validity. If an unnamed FLQuant object is provided, this is used for sizing but not stored in any slot.

Methods

Methods exist for various calculations based on values stored in the class:

ccost

Calculate the total crew costs by evaluating the formula in crewshare.

cost

Total costs, calculated.

lrevenue

.

revenue

.

profit

.

landings

.

discards

.

catch

.

catch.n

.

catch.wt

.

harvest

.

See also

Author

Iago Mosqueira (EC JRC) iago.mosqueira@ec.europa.eu

Maintainer: Iago Mosqueira (EC JRC) iago.mosqueira@ec.europa.eu

Iago Mosqueira, EC JRC.

Examples

if (FALSE) FLFishery()

data(ple4)
FLFishery(PLE=as(ple4, 'FLCatch'))
#> An object of class "FLFishery"
#> 
#> Name:  
#> Description:  
#> Names :  PLE 
#> Quant: quant 
#> Dims:  quant 	year	unit	season	area	iter
#> 	2	61	1	1	1	1	
#> 
#> Range:  min	max	minyear	maxyear 
#> 	0	0	1	1	
#> 
#> crewshare     An object of class "predictModel":
#> ~ fixed * crew + share * lrevenue 
#> params:
#>   fixed, cre  : [ 3 1 ], units =  EUR