An internal class for homogeneous fishing fleets

The same as the FLFishery class but all predictModel slots have been turned into FLQuant objects. The class is used for passing to C++ FLR objects.

# S4 method for FLFisherycpp
capacity(object)

# S4 method for FLFisherycpp,FLQuant
capacity(object) <- value

# S4 method for FLFisherycpp,numeric
capacity(object) <- value

# S4 method for FLFisherycpp
hperiod(object)

# S4 method for FLFisherycpp,FLQuant
hperiod(object) <- value

# S4 method for FLFisherycpp,numeric
hperiod(object) <- value

# S4 method for FLFisherycpp
orevenue(object)

# S4 method for FLFisherycpp,FLQuant
orevenue(object) <- value

# S4 method for FLFisherycpp,numeric
orevenue(object) <- value

# S4 method for FLFisherycpp,ANY
effort(object)

# S4 method for FLFisherycpp,FLQuant
effort(object) <- value

# S4 method for FLFisherycpp,numeric
effort(object) <- value

# S4 method for FLFisherycpp
vcost(object)

# S4 method for FLFisherycpp,FLQuant
vcost(object) <- value

# S4 method for FLFisherycpp,numeric
vcost(object) <- value

# S4 method for FLFisherycpp
fcost(object)

# S4 method for FLFisherycpp,FLQuant
fcost(object) <- value

# S4 method for FLFisherycpp,numeric
fcost(object) <- value

# S4 method for FLFisherycpp
crewshare(object)

# S4 method for FLFisherycpp,FLQuant
crewshare(object) <- value

# S4 method for FLFisherycpp,numeric
crewshare(object) <- value

Arguments

object
Object to extract or operate on

value
Replacement value

Details

What have you done what's in your mind what do you need Where shall we go to let it out What have you seen, we don't know where you've been Life so often blows your candle out Believe in what is right, what's right for you tonight Who knows what the fuck it's all about

Slots

FLFisherycpp objects inherit from FLCatches woth a number of slots added.

.Data

The list of FLCatch object with catch data per stock, (FLCatches).

name

Name of the object, e.g. species or stock code, (character).

desc

Description of the data contents and origin, (character).

range

Ranges of age and years, plusgroup, (numeric).

capacity

Number of boats in the fleet, (FLQuant).

effort

Mean effort per boat applied by the fleet, (FLQuant).

hperiod

Start and end of fishing within each time step, as proportions. An FLQuant object with dimnames `quant=c('start', 'end')` in the first dimension.

vcost

Variable costs per unit of effort, (FLQuant).

fcost

Variable costs per unit of effort, (FLQuant).

orevenue

Revenues obtained from sources other than landings, (FLQuant).

crewshare

Formula, parameter values and inputs to calculate the crew costs, (FLQuant).

Validity

You can inspect the class validity function by using getValidity(getClassDef('FLFishery'))

Accessors

All slots in the class have accessor and replacement methods defined that allow retrieving and substituting individual slots.

The values passed for replacement need to be of the class of that slot. A numeric vector can also be used when replacing FLQuant slots, and the vector will be used to substitute the values in the slot, but not its other attributes.

Constructor

A construction method exists for this class that can take named arguments for any of its slots. All slots are then created to match the requirements of the class validity. If an unnamed FLQuant object is provided, this is used for sizing but not stored in any slot.

Methods

Methods exist for various calculations based on values stored in the class:

ccost

Calculate the total crew costs by evaluating the formula in crewshare.

cost

Total costs, calculated.

lrevenue

.

revenue

.

profit

.

landings

.

discards

.

catch

.

catch.n

.

catch.wt

.

harvest

.

See also

Examples

data(ple4)
FLFishery(PLE=as(ple4, 'FLCatch'))

#> An object of class "FLFishery"
#> 
#> Name:  
#> Description:  
#> Names :  PLE 
#> Quant: quant 
#> Dims:  quant 	year	unit	season	area	iter
#> 	2	61	1	1	1	1	
#> 
#> Range:  min	max	minyear	maxyear 
#> 	0	0	1	1	
#> 
#> capacity      : [ 1 61 1 1 1 1 ], units =  NA 
#> effort        : [ 1 61 1 1 1 1 ], units =   
#> hperiod       : [ 2 61 1 1 1 1 ], units =   
#> vcost         : [ 1 61 1 1 1 1 ], units =   
#> fcost         : [ 1 61 1 1 1 1 ], units =   
#> orevenue      : [ 1 61 1 1 1 1 ], units =   
#> crewshare     An object of class "predictModel":
#> ~ fixed * crew + share * lrevenue 
#> params:
#>   fixed, cre  : [ 3 1 ], units =  EUR