An internal class for homogeneous fishing fleets

The same as the FLFishery class but all predictModel slots have been turned into FLQuant objects. The class is used for passing to C++ FLR objects.

# S4 method for FLFisherycpp
capacity(object)

# S4 method for FLFisherycpp,FLQuant
capacity(object) <- value

# S4 method for FLFisherycpp,numeric
capacity(object) <- value

# S4 method for FLFisherycpp
hperiod(object)

# S4 method for FLFisherycpp,FLQuant
hperiod(object) <- value

# S4 method for FLFisherycpp,numeric
hperiod(object) <- value

# S4 method for FLFisherycpp
orevenue(object)

# S4 method for FLFisherycpp,FLQuant
orevenue(object) <- value

# S4 method for FLFisherycpp,numeric
orevenue(object) <- value

# S4 method for FLFisherycpp,ANY
effort(object)

# S4 method for FLFisherycpp,FLQuant
effort(object) <- value

# S4 method for FLFisherycpp,numeric
effort(object) <- value

# S4 method for FLFisherycpp
vcost(object)

# S4 method for FLFisherycpp,FLQuant
vcost(object) <- value

# S4 method for FLFisherycpp,numeric
vcost(object) <- value

# S4 method for FLFisherycpp
fcost(object)

# S4 method for FLFisherycpp,FLQuant
fcost(object) <- value

# S4 method for FLFisherycpp,numeric
fcost(object) <- value

# S4 method for FLFisherycpp
crewshare(object)

# S4 method for FLFisherycpp,FLQuant
crewshare(object) <- value

# S4 method for FLFisherycpp,numeric
crewshare(object) <- value

Arguments

object
Object to extract or operate on

value
Replacement value

Details

What have you done what's in your mind what do you need Where shall we go to let it out What have you seen, we don't know where you've been Life so often blows your candle out Believe in what is right, what's right for you tonight Who knows what the fuck it's all about

Slots

FLFisherycpp objects inherit from FLCatches woth a number of slots added.

.Data

The list of FLCatch object with catch data per stock, (FLCatches).

name

Name of the object, e.g. species or stock code, (character).

desc

Description of the data contents and origin, (character).

range

Ranges of age and years, plusgroup, (numeric).

capacity

Number of boats in the fleet, (FLQuant).

effort

Mean effort per boat applied by the fleet, (FLQuant).

hperiod

Start and end of fishing within each time step, as proportions. An FLQuant object with dimnames `quant=c('start', 'end')` in the first dimension.

vcost

Variable costs per unit of effort, (FLQuant).

fcost

Variable costs per unit of effort, (FLQuant).

orevenue

Revenues obtained from sources other than landings, (FLQuant).

crewshare

Formula, parameter values and inputs to calculate the crew costs, (FLQuant).

Validity

You can inspect the class validity function by using getValidity(getClassDef('FLFishery'))

Accessors

All slots in the class have accessor and replacement methods defined that allow retrieving and substituting individual slots.

The values passed for replacement need to be of the class of that slot. A numeric vector can also be used when replacing FLQuant slots, and the vector will be used to substitute the values in the slot, but not its other attributes.

Constructor

A construction method exists for this class that can take named arguments for any of its slots. All slots are then created to match the requirements of the class validity. If an unnamed FLQuant object is provided, this is used for sizing but not stored in any slot.

Methods

Methods exist for various calculations based on values stored in the class:

ccost

Calculate the total crew costs by evaluating the formula in crewshare.

cost

Total costs, calculated.

lrevenue

.

revenue

.

profit

.

landings

.

discards

.

catch

.

catch.n

.

catch.wt

.

harvest

.

See also

Examples

data(ple4)
FLFishery(PLE=as(ple4, 'FLCatch'))

#> An object of class "FLFishery"
#> [[1]]
#> An object of class "FLCatch"
#> Slot "landings.n":
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  1957       1958       1959       1960       1961       1962      
#>   1  6.5789e-01 5.4770e-01 4.2939e-01 6.5634e-01 6.9398e-01 6.3998e-01
#>   2  3.8177e+03 6.6619e+03 1.4099e+04 5.2745e+03 2.1137e+03 1.6658e+03
#>   3  6.9681e+04 2.5992e+04 3.2342e+04 6.7734e+04 3.7875e+04 3.6859e+04
#>   4  5.5543e+04 6.4841e+04 2.6260e+04 5.0500e+04 7.1683e+04 6.6631e+04
#>   5  3.1042e+04 3.3854e+04 3.8955e+04 2.2867e+04 3.0956e+04 4.5568e+04
#>   6  1.0960e+04 2.1760e+04 2.3703e+04 2.6810e+04 1.4596e+04 2.0406e+04
#>   7  1.2049e+04 8.5678e+03 1.4322e+04 1.3940e+04 1.5632e+04 8.7376e+03
#>   8  9.5953e+03 8.6104e+03 5.9994e+03 9.7186e+03 9.0449e+03 9.7869e+03
#>   9  5.4948e+03 7.0501e+03 6.9324e+03 4.5918e+03 6.5316e+03 5.7697e+03
#>   10 1.1779e+04 1.4773e+04 1.6725e+04 1.5503e+04 1.2505e+04 1.3593e+04
#>     year
#> age  1963       1964       1965       1966       1967       1968      
#>   1  2.8480e-01 1.2662e+00 4.7605e-01 5.1361e-01 9.3693e-01 3.5445e-01
#>   2  3.5115e+03 1.0302e+04 1.0165e+04 2.8034e+03 4.2952e+03 9.8082e+03
#>   3  2.3301e+04 4.9070e+04 5.1526e+04 7.3702e+04 3.1491e+04 4.9859e+04
#>   4  7.4463e+04 5.9743e+04 5.7700e+04 4.8162e+04 1.1185e+05 4.2741e+04
#>   5  5.5803e+04 5.0215e+04 5.2662e+04 3.3914e+04 3.5064e+04 1.3382e+05
#>   6  3.0795e+04 3.8102e+04 3.2134e+04 3.0584e+04 1.9956e+04 2.1860e+04
#>   7  1.2304e+04 1.7096e+04 1.8841e+04 1.5318e+04 1.5625e+04 1.1270e+04
#>   8  5.3718e+03 7.2567e+03 9.3691e+03 1.0118e+04 8.7300e+03 9.4646e+03
#>   9  6.7091e+03 3.8234e+03 4.8058e+03 5.8320e+03 6.3968e+03 5.7308e+03
#>   10 1.6653e+04 1.9738e+04 1.6512e+04 1.3434e+04 1.2784e+04 1.4154e+04
#>     year
#> age  1969       1970       1971       1972       1973       1974      
#>   1  4.7193e+00 7.1123e+01 2.1241e+01 1.9401e+03 2.4524e+03 2.3357e+03
#>   2  1.6733e+04 1.3334e+04 3.0426e+04 3.3804e+04 3.0846e+04 2.2257e+04
#>   3  5.5281e+04 4.2623e+04 5.4528e+04 7.0831e+04 7.2701e+04 6.2826e+04
#>   4  3.9635e+04 4.0541e+04 3.6938e+04 6.1244e+04 7.3652e+04 4.7563e+04
#>   5  4.4853e+04 3.5074e+04 2.7324e+04 2.5898e+04 4.1280e+04 4.0036e+04
#>   6  1.0146e+05 3.0042e+04 2.5504e+04 1.7726e+04 1.5987e+04 2.2241e+04
#>   7  1.2474e+04 5.7327e+04 1.7180e+04 1.5407e+04 1.0556e+04 8.5997e+03
#>   8  6.7471e+03 7.2149e+03 3.3998e+04 1.0747e+04 9.7098e+03 6.3599e+03
#>   9  6.1857e+03 4.3645e+03 4.7675e+03 2.3096e+04 7.3212e+03 6.4546e+03
#>   10 1.5527e+04 1.7101e+04 1.6488e+04 1.5528e+04 2.6610e+04 2.2141e+04
#>     year
#> age  1975       1976       1977       1978       1979       1980      
#>   1  7.9533e+02 2.8953e+03 3.8608e+03 1.2587e+03 1.0612e+03 9.9989e+02
#>   2  2.2860e+04 3.3043e+04 5.0444e+04 6.1272e+04 4.2082e+04 7.6672e+04
#>   3  5.9487e+04 8.6033e+04 4.0060e+04 6.6284e+04 1.0279e+05 1.3879e+05
#>   4  3.2140e+04 1.0852e+05 6.4320e+04 5.7828e+04 4.1028e+04 7.9631e+04
#>   5  2.2787e+04 1.7696e+04 6.3651e+04 5.6708e+04 3.9916e+04 2.8611e+04
#>   6  1.8797e+04 1.1049e+04 9.8050e+03 4.1042e+04 3.1656e+04 1.9781e+04
#>   7  1.1126e+04 9.9208e+03 6.3543e+03 5.5722e+03 2.1132e+04 1.5387e+04
#>   8  4.9979e+03 6.5284e+03 5.8463e+03 3.5865e+03 2.9882e+03 1.1113e+04
#>   9  4.0785e+03 3.0881e+03 3.8608e+03 3.3403e+03 2.0283e+03 1.6764e+03
#>   10 1.7837e+04 1.3249e+04 9.7368e+03 8.3180e+03 7.6179e+03 6.6988e+03
#>     year
#> age  1981       1982       1983       1984       1985       1986      
#>   1  3.5815e+02 3.0821e+03 1.4442e+03 1.1534e+02 1.3957e+02 1.6157e+03
#>   2  9.3948e+04 4.7712e+04 9.9627e+04 5.3064e+04 7.1096e+04 5.0046e+04
#>   3  1.5670e+05 1.9098e+05 1.4050e+05 2.7830e+05 1.8277e+05 1.6520e+05
#>   4  6.4677e+04 6.8562e+04 8.9117e+04 6.7196e+04 1.8454e+05 1.0282e+05
#>   5  3.5504e+04 2.6999e+04 2.8826e+04 4.0110e+04 4.5802e+04 9.4994e+04
#>   6  1.3468e+04 1.6686e+04 1.2668e+04 1.4187e+04 2.0635e+04 2.2913e+04
#>   7  9.9944e+03 7.0638e+03 8.8640e+03 6.8932e+03 7.6803e+03 1.0370e+04
#>   8  8.0743e+03 5.2637e+03 3.9297e+03 5.3125e+03 4.1377e+03 4.2546e+03
#>   9  6.1060e+03 4.5007e+03 3.1720e+03 2.4856e+03 3.2548e+03 2.4372e+03
#>   10 5.8858e+03 8.1906e+03 8.2328e+03 6.9825e+03 5.8567e+03 6.2992e+03
#>     year
#> age  1987       1988       1989       1990       1991       1992      
#>   1  1.8335e+00 1.6052e+00 1.4339e+03 2.1625e+03 1.4514e+03 4.0036e+03
#>   2  6.1107e+04 1.1681e+04 3.8555e+04 3.2202e+04 3.2331e+04 4.0289e+04
#>   3  1.1953e+05 2.6573e+05 1.1904e+05 1.1790e+05 8.5073e+04 9.6266e+04
#>   4  1.1338e+05 8.6239e+04 2.4853e+05 1.2362e+05 1.1510e+05 8.1455e+04
#>   5  6.4848e+04 5.5601e+04 6.3410e+04 1.6398e+05 7.5259e+04 7.8895e+04
#>   6  4.6403e+04 3.0955e+04 2.6254e+04 3.2831e+04 7.7313e+04 3.7446e+04
#>   7  1.0445e+04 1.9869e+04 1.2392e+04 9.8274e+03 1.2398e+04 3.1918e+04
#>   8  5.2475e+03 4.8810e+03 8.4475e+03 5.1816e+03 4.6301e+03 6.5467e+03
#>   9  2.5027e+03 2.9164e+03 2.4119e+03 4.2145e+03 3.0603e+03 3.0239e+03
#>   10 6.6599e+03 6.3268e+03 5.1187e+03 3.9767e+03 5.1378e+03 5.8602e+03
#>     year
#> age  1993       1994       1995       1996       1997       1998      
#>   1  3.5826e+03 2.0202e+03 7.7925e+03 1.0332e+03 1.8863e+03 1.7248e+02
#>   2  5.4549e+04 4.0776e+04 3.9586e+04 2.6907e+04 1.9919e+04 3.1729e+04
#>   3  1.2357e+05 1.1668e+05 8.9268e+04 9.1876e+04 7.6088e+04 6.2021e+04
#>   4  8.7868e+04 8.7922e+04 7.8919e+04 5.2213e+04 5.8220e+04 4.6221e+04
#>   5  5.9429e+04 5.0755e+04 4.4154e+04 3.8411e+04 2.5130e+04 2.4023e+04
#>   6  4.0442e+04 3.0042e+04 2.4930e+04 2.1364e+04 1.8403e+04 1.1394e+04
#>   7  1.6523e+04 1.7190e+04 1.2590e+04 1.0906e+04 9.4345e+03 7.2080e+03
#>   8  1.5502e+04 6.6012e+03 6.5862e+03 5.4044e+03 4.9141e+03 3.7540e+03
#>   9  3.5693e+03 6.3012e+03 2.5419e+03 2.9527e+03 2.6443e+03 2.2355e+03
#>   10 5.5121e+03 4.3702e+03 4.9901e+03 4.1743e+03 4.4094e+03 4.0650e+03
#>     year
#> age  1999       2000       2001       2002       2003       2004      
#>   1  3.9252e+02 3.1548e+03 1.2647e+04 1.0787e+03 9.3859e+02 5.4785e+02
#>   2  7.4839e+03 1.1602e+04 1.7878e+04 1.4064e+04 4.1938e+04 9.0527e+03
#>   3  2.0489e+05 6.5388e+04 3.4016e+04 8.6682e+04 6.4293e+04 1.5008e+05
#>   4  3.5388e+04 1.7580e+05 4.0149e+04 6.2282e+04 5.4604e+04 3.3910e+04
#>   5  3.4317e+04 3.1336e+04 7.1644e+04 3.9085e+04 3.0207e+04 3.0249e+04
#>   6  9.7643e+03 1.2401e+04 1.2092e+04 4.6403e+04 1.9650e+04 1.5722e+04
#>   7  3.7197e+03 2.8281e+03 4.6526e+03 5.1624e+03 1.4683e+04 7.1620e+03
#>   8  2.5269e+03 1.2493e+03 1.4536e+03 2.1610e+03 1.8182e+03 5.5146e+03
#>   9  1.6230e+03 1.2092e+03 7.1574e+02 5.9306e+02 6.7339e+02 5.4110e+02
#>   10 3.1986e+03 2.3946e+03 1.8995e+03 1.5026e+03 1.3239e+03 1.3297e+03
#>     year
#> age  2005       2006       2007       2008       2009       2010      
#>   1  5.2488e+03 2.4356e+02 3.4668e+03 4.5836e+02 1.7398e+03 2.5969e+03
#>   2  1.4580e+04 1.4330e+04 1.6356e+04 1.2723e+04 1.0729e+04 1.4038e+04
#>   3  3.2935e+04 7.4279e+04 4.6475e+04 6.9598e+04 7.8334e+04 5.6436e+04
#>   4  9.3409e+04 2.9512e+04 6.3599e+04 4.6508e+04 4.6007e+04 7.6859e+04
#>   5  1.5802e+04 5.0303e+04 1.5763e+04 3.5005e+04 2.3614e+04 2.6751e+04
#>   6  9.4463e+03 8.4925e+03 1.9966e+04 8.7677e+03 1.8686e+04 1.2075e+04
#>   7  5.5446e+03 5.4649e+03 4.1308e+03 6.8550e+03 4.9948e+03 9.5416e+03
#>   8  2.2982e+03 2.1221e+03 1.4459e+03 9.4816e+02 3.3813e+03 1.9635e+03
#>   9  1.9018e+03 8.2947e+02 7.3777e+02 7.7574e+02 5.6297e+02 2.1486e+03
#>   10 1.1816e+03 1.7433e+03 1.4685e+03 1.4752e+03 1.7861e+03 2.0284e+03
#>     year
#> age  2011       2012       2013       2014       2015       2016      
#>   1  5.6268e+02 1.7458e+02 5.5317e+02 5.3145e+00 7.6366e-01 8.1135e-01
#>   2  1.3669e+04 3.9728e+03 1.1141e+04 5.7111e+03 2.9733e+03 3.1913e+03
#>   3  5.2667e+04 6.0547e+04 7.2666e+04 4.3317e+04 4.4096e+04 4.4511e+04
#>   4  5.7594e+04 5.8466e+04 9.0734e+04 8.2026e+04 7.2276e+04 8.0027e+04
#>   5  4.6736e+04 4.6368e+04 4.8886e+04 5.0832e+04 6.5415e+04 4.8613e+04
#>   6  1.4055e+04 3.2279e+04 2.9393e+04 2.4847e+04 3.3683e+04 3.6922e+04
#>   7  7.5574e+03 8.1146e+03 1.5433e+04 1.4507e+04 1.4591e+04 1.7826e+04
#>   8  3.9503e+03 3.6327e+03 4.2936e+03 8.4909e+03 7.6212e+03 5.5445e+03
#>   9  9.8736e+02 2.7629e+03 2.1229e+03 2.3395e+03 3.8729e+03 3.0372e+03
#>   10 3.6733e+03 4.0976e+03 5.9574e+03 6.8204e+03 7.3928e+03 8.7017e+03
#>     year
#> age  2017      
#>   1  2.4236e+01
#>   2  5.0710e+03
#>   3  2.6847e+04
#>   4  6.5405e+04
#>   5  5.9136e+04
#>   6  3.4228e+04
#>   7  1.9948e+04
#>   8  7.6049e+03
#>   9  2.5668e+03
#>   10 8.8813e+03
#> 
#> units:  1000 
#> 
#> Slot "landings.wt":
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  1957  1958  1959  1960  1961  1962  1963  1964  1965  1966  1967  1968 
#>   1  0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#>   2  0.165 0.198 0.218 0.200 0.191 0.211 0.253 0.250 0.242 0.232 0.232 0.267
#>   3  0.201 0.221 0.246 0.236 0.233 0.248 0.286 0.273 0.282 0.270 0.279 0.298
#>   4  0.258 0.259 0.293 0.289 0.302 0.300 0.319 0.312 0.321 0.348 0.322 0.331
#>   5  0.353 0.337 0.362 0.386 0.412 0.400 0.399 0.388 0.385 0.436 0.425 0.366
#>   6  0.456 0.453 0.473 0.485 0.509 0.541 0.533 0.487 0.471 0.484 0.547 0.517
#>   7  0.533 0.513 0.592 0.601 0.604 0.570 0.624 0.628 0.539 0.559 0.597 0.590
#>   8  0.589 0.615 0.623 0.683 0.671 0.692 0.667 0.700 0.663 0.624 0.662 0.596
#>   9  0.396 0.665 0.750 0.724 0.812 0.777 0.715 0.737 0.726 0.690 0.738 0.686
#>   10 0.998 0.992 1.000 1.094 1.071 1.127 1.028 1.005 0.887 0.933 0.978 0.911
#>     year
#> age  1969  1970  1971  1972  1973  1974  1975  1976  1977  1978  1979  1980 
#>   1  0.217 0.315 0.256 0.246 0.272 0.285 0.249 0.265 0.254 0.244 0.235 0.238
#>   2  0.294 0.286 0.318 0.296 0.316 0.311 0.300 0.295 0.323 0.315 0.311 0.286
#>   3  0.310 0.318 0.356 0.352 0.344 0.354 0.330 0.338 0.353 0.369 0.349 0.344
#>   4  0.333 0.356 0.403 0.428 0.405 0.405 0.420 0.375 0.380 0.397 0.388 0.401
#>   5  0.359 0.419 0.448 0.493 0.486 0.476 0.495 0.513 0.418 0.438 0.429 0.473
#>   6  0.412 0.443 0.514 0.541 0.539 0.554 0.587 0.594 0.556 0.491 0.474 0.545
#>   7  0.573 0.499 0.542 0.608 0.605 0.609 0.636 0.641 0.647 0.609 0.550 0.588
#>   8  0.655 0.672 0.607 0.646 0.627 0.693 0.703 0.705 0.721 0.687 0.675 0.662
#>   9  0.658 0.744 0.699 0.674 0.677 0.707 0.783 0.741 0.715 0.776 0.796 0.772
#>   10 0.893 0.892 0.891 0.939 0.842 0.926 1.019 0.980 0.978 0.950 0.960 1.013
#>     year
#> age  1981  1982  1983  1984  1985  1986  1987  1988  1989  1990  1991  1992 
#>   1  0.237 0.279 0.200 0.231 0.245 0.221 0.000 0.000 0.236 0.271 0.227 0.251
#>   2  0.274 0.262 0.250 0.263 0.264 0.269 0.249 0.254 0.280 0.284 0.286 0.263
#>   3  0.329 0.311 0.300 0.283 0.290 0.303 0.299 0.278 0.308 0.297 0.292 0.290
#>   4  0.416 0.424 0.383 0.364 0.335 0.339 0.345 0.341 0.331 0.315 0.302 0.312
#>   5  0.505 0.514 0.515 0.480 0.445 0.405 0.378 0.418 0.385 0.364 0.360 0.330
#>   6  0.558 0.608 0.604 0.591 0.563 0.473 0.472 0.478 0.515 0.441 0.452 0.415
#>   7  0.604 0.664 0.677 0.677 0.667 0.668 0.574 0.590 0.591 0.586 0.526 0.530
#>   8  0.642 0.712 0.771 0.726 0.730 0.750 0.728 0.680 0.668 0.690 0.666 0.607
#>   9  0.725 0.738 0.815 0.839 0.807 0.856 0.835 0.808 0.785 0.761 0.743 0.719
#>   10 1.007 0.984 0.984 1.036 1.021 1.014 0.993 1.017 0.940 1.010 0.924 0.891
#>     year
#> age  1993  1994  1995  1996  1997  1998  1999  2000  2001  2002  2003  2004 
#>   1  0.249 0.229 0.272 0.240 0.208 0.151 0.245 0.228 0.238 0.237 0.232 0.214
#>   2  0.273 0.263 0.277 0.279 0.271 0.260 0.253 0.267 0.267 0.264 0.252 0.246
#>   3  0.288 0.284 0.301 0.304 0.313 0.306 0.280 0.283 0.291 0.289 0.285 0.281
#>   4  0.319 0.333 0.335 0.346 0.355 0.384 0.347 0.312 0.307 0.311 0.320 0.328
#>   5  0.343 0.375 0.375 0.415 0.410 0.452 0.415 0.378 0.360 0.336 0.353 0.391
#>   6  0.408 0.417 0.420 0.465 0.474 0.546 0.416 0.461 0.412 0.430 0.389 0.429
#>   7  0.512 0.491 0.474 0.490 0.541 0.613 0.538 0.597 0.582 0.477 0.482 0.508
#>   8  0.630 0.610 0.593 0.553 0.574 0.673 0.637 0.689 0.701 0.644 0.635 0.560
#>   9  0.720 0.731 0.734 0.712 0.616 0.687 0.748 0.752 0.796 0.760 0.763 0.797
#>   10 0.856 0.906 0.906 0.858 0.912 0.899 0.804 0.888 0.799 0.904 0.857 0.872
#>     year
#> age  2005  2006  2007  2008  2009  2010  2011  2012  2013  2014  2015  2016 
#>   1  0.272 0.253 0.263 0.249 0.176 0.206 0.235 0.236 0.031 0.207 0.000 0.000
#>   2  0.265 0.267 0.268 0.269 0.260 0.265 0.242 0.258 0.242 0.252 0.251 0.249
#>   3  0.280 0.282 0.303 0.309 0.308 0.308 0.281 0.305 0.281 0.285 0.284 0.271
#>   4  0.330 0.322 0.343 0.341 0.355 0.348 0.341 0.351 0.313 0.318 0.321 0.296
#>   5  0.382 0.383 0.364 0.400 0.415 0.418 0.414 0.380 0.364 0.368 0.359 0.350
#>   6  0.426 0.389 0.432 0.446 0.481 0.476 0.504 0.436 0.417 0.418 0.409 0.385
#>   7  0.465 0.457 0.507 0.531 0.531 0.516 0.604 0.518 0.494 0.479 0.473 0.450
#>   8  0.555 0.477 0.486 0.720 0.608 0.625 0.521 0.558 0.600 0.543 0.487 0.531
#>   9  0.617 0.531 0.587 0.640 0.668 0.682 0.556 0.558 0.607 0.628 0.582 0.556
#>   10 0.910 0.748 0.632 0.638 0.792 0.649 0.804 0.680 0.680 0.650 0.600 0.684
#>     year
#> age  2017 
#>   1  0.212
#>   2  0.247
#>   3  0.276
#>   4  0.299
#>   5  0.357
#>   6  0.410
#>   7  0.455
#>   8  0.543
#>   9  0.642
#>   10 0.735
#> 
#> units:  kg 
#> 
#> Slot "discards.n":
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  1957       1958       1959       1960       1961       1962      
#>   1  4.2702e+04 7.2732e+04 9.9992e+04 9.7351e+04 1.0522e+05 6.5815e+04
#>   2  3.6324e+04 6.4633e+04 1.0650e+05 1.4762e+05 1.3291e+05 1.4551e+05
#>   3  9.6720e+03 2.6039e+04 4.8277e+04 4.8880e+04 9.9414e+04 9.3109e+04
#>   4  1.0172e+03 2.5281e+03 1.1504e+04 9.8080e+02 5.5972e+02 2.2417e+04
#>   5  8.4596e+02 1.9949e+03 4.3953e+03 1.1374e+03 1.5031e+03 7.5291e+02
#>   6  2.8323e+01 6.8403e+01 1.0773e+02 5.0627e+02 1.4699e+02 1.9087e+02
#>   7  4.9188e-01 4.9258e-01 5.0523e-01 4.8810e-01 5.2423e-01 4.9885e-01
#>   8  4.7852e-01 5.0605e-01 4.5817e-01 5.3558e-01 4.5679e-01 5.6030e-01
#>   9  4.9744e-01 5.4251e-01 5.1434e-01 4.4832e-01 5.3626e-01 5.6768e-01
#>   10 4.8836e-01 5.2829e-01 5.0589e-01 5.0817e-01 4.1993e-01 5.4928e-01
#>     year
#> age  1963       1964       1965       1966       1967       1968      
#>   1  5.1939e+04 1.6773e+05 4.1740e+04 3.9663e+04 3.7965e+04 5.2596e+04
#>   2  1.0809e+05 1.0656e+05 4.3682e+05 1.0836e+05 9.4015e+04 7.2620e+04
#>   3  1.3817e+05 7.7181e+04 7.1952e+04 3.8897e+05 8.9191e+04 5.5163e+04
#>   4  9.4626e+03 3.0985e+04 3.6727e+03 1.3228e+04 1.4812e+05 2.8889e+04
#>   5  2.1273e+03 2.6146e+03 1.1073e+03 1.2739e+03 5.1802e+02 2.3601e+04
#>   6  1.7908e+02 3.2984e+02 2.8350e+02 1.3772e+02 2.3113e+02 7.2501e+01
#>   7  4.6975e-01 5.3633e-01 5.1021e-01 4.5586e-01 4.7359e-01 5.2948e-01
#>   8  3.8560e-01 5.4604e-01 5.4456e-01 4.6067e-01 5.3471e-01 4.5517e-01
#>   9  4.6697e-01 4.7670e-01 5.6711e-01 4.4278e-01 4.6611e-01 7.0929e-01
#>   10 4.9231e-01 5.8009e-01 4.6680e-01 4.4187e-01 5.6085e-01 5.1451e-01
#>     year
#> age  1969       1970       1971       1972       1973       1974      
#>   1  1.0945e+05 1.1789e+05 7.3180e+04 5.9135e+04 2.4354e+05 2.1392e+05
#>   2  7.0945e+04 1.2066e+05 9.6519e+04 4.9921e+04 4.6086e+04 2.8634e+05
#>   3  1.9361e+04 2.3720e+04 4.6305e+04 3.7542e+04 4.6931e+03 3.9842e+03
#>   4  2.0068e+04 9.5401e+02 2.8579e+03 6.4331e+03 6.4614e+02 3.1090e+02
#>   5  9.2130e+02 4.4782e+03 4.3290e+02 6.2025e+02 1.1989e+03 5.7485e+02
#>   6  1.9026e+03 1.4197e+02 8.7853e+01 7.4540e+01 7.4329e+01 1.1943e+02
#>   7  5.6613e-01 3.7292e-01 4.9165e-01 5.2095e-01 5.6534e-01 5.1803e-01
#>   8  5.2376e-01 5.8662e-01 5.5350e-01 4.7925e-01 4.6917e-01 4.8516e-01
#>   9  3.5675e-01 5.3511e-01 3.5975e-01 6.0426e-01 5.7260e-01 5.5897e-01
#>   10 4.5588e-01 5.9132e-01 4.7192e-01 4.8242e-01 5.2811e-01 4.8576e-01
#>     year
#> age  1975       1976       1977       1978       1979       1980      
#>   1  1.8686e+05 1.8408e+05 3.0715e+05 2.5898e+05 2.3636e+05 2.2705e+05
#>   2  2.1433e+05 1.3498e+05 9.1669e+04 1.6166e+05 1.5624e+05 1.4218e+05
#>   3  1.7314e+05 7.6776e+04 7.3282e+04 3.0055e+04 4.9878e+04 9.4972e+02
#>   4  4.6524e+03 1.9777e+04 3.2639e+04 1.1892e+04 1.5430e+04 6.9268e+02
#>   5  2.1391e+02 2.1855e+02 7.0737e+03 1.4608e+03 4.9221e+02 2.1669e+02
#>   6  1.0313e+02 4.5149e+01 1.3866e+02 6.8383e+02 2.4656e+02 9.3389e+01
#>   7  4.4101e-01 5.3040e-01 5.3649e-01 4.8640e-01 4.3684e-01 5.4882e-01
#>   8  3.9711e-01 4.9806e-01 5.8107e-01 5.5473e-01 3.5963e-01 6.4452e-01
#>   9  4.6745e-01 5.1088e-01 4.7452e-01 7.1589e-01 3.6184e-01 4.4146e-01
#>   10 4.2625e-01 5.1468e-01 5.2881e-01 5.7752e-01 4.1077e-01 6.0925e-01
#>     year
#> age  1981       1982       1983       1984       1985       1986      
#>   1  1.8514e+05 3.7578e+05 3.1094e+05 3.3021e+05 4.6798e+05 1.0816e+06
#>   2  1.7566e+05 2.0186e+05 3.6317e+05 2.6194e+05 2.2382e+05 3.6709e+05
#>   3  2.3026e+03 4.1624e+03 3.7418e+04 5.2242e+04 4.4268e+04 4.8280e+04
#>   4  4.0555e+02 1.0689e+03 1.9828e+03 2.6232e+04 2.0332e+03 2.5565e+04
#>   5  4.1777e+02 2.0024e+02 7.8368e+02 1.4428e+03 2.4205e+02 1.3560e+03
#>   6  5.8620e+01 9.6820e+01 7.0866e+01 6.3955e+01 9.4690e+01 1.2001e+02
#>   7  5.0963e-01 4.5841e-01 5.3795e-01 5.0373e-01 5.1602e-01 5.2699e-01
#>   8  4.8666e-01 4.7406e-01 4.6515e-01 4.7487e-01 5.4725e-01 4.5221e-01
#>   9  6.0994e-01 4.9567e-01 5.2630e-01 4.5531e-01 5.7848e-01 4.2992e-01
#>   10 4.8311e-01 4.6665e-01 4.9543e-01 5.3287e-01 5.0226e-01 4.4467e-01
#>     year
#> age  1987       1988       1989       1990       1991       1992      
#>   1  4.4268e+05 3.7416e+05 2.4375e+05 2.0322e+05 1.8651e+05 1.6404e+05
#>   2  9.8615e+05 4.6730e+05 4.0016e+05 2.6636e+05 2.2641e+05 2.0216e+05
#>   3  1.8266e+05 4.8095e+05 2.1239e+05 1.8977e+05 1.3812e+05 1.0543e+05
#>   4  1.3401e+03 6.1806e+04 8.9460e+04 2.8992e+04 3.8363e+04 3.4342e+04
#>   5  1.0755e+03 8.1848e+02 7.7953e+03 4.8491e+03 5.4692e+03 3.8385e+03
#>   6  2.5861e+02 1.7447e+02 1.3995e+02 1.9460e+02 9.2261e+02 8.7822e+02
#>   7  3.8572e-01 5.1348e-01 5.3389e-01 5.4276e-01 4.0226e-01 5.2421e-01
#>   8  5.6195e-01 3.9370e-01 5.2629e-01 5.5243e-01 4.2944e-01 4.5279e-01
#>   9  5.3324e-01 4.2461e-01 5.1957e-01 5.5399e-01 4.7594e-01 4.5883e-01
#>   10 5.9348e-01 4.9410e-01 4.4078e-01 4.7270e-01 4.5593e-01 4.2002e-01
#>     year
#> age  1993       1994       1995       1996       1997       1998      
#>   1  1.0105e+05 9.1270e+04 1.1906e+05 1.0416e+05 2.7031e+05 9.1485e+04
#>   2  1.5694e+05 8.7587e+04 8.9031e+04 2.0551e+05 2.3600e+05 6.7022e+05
#>   3  5.9538e+04 3.9351e+04 1.7015e+04 3.7214e+04 1.6629e+05 1.5966e+05
#>   4  1.3004e+04 1.0496e+03 8.0103e+02 3.8750e+03 6.9410e+02 3.9559e+04
#>   5  1.5361e+03 8.7744e+02 6.1145e+02 4.6513e+02 2.1488e+02 3.1354e+02
#>   6  2.4986e+02 1.5683e+02 1.2225e+02 1.2990e+02 1.0693e+02 4.1709e+01
#>   7  5.6027e-01 4.9740e-01 4.2687e-01 4.3976e-01 4.9561e-01 6.3189e-01
#>   8  6.7395e-01 4.9548e-01 6.2301e-01 4.5941e-01 4.8774e-01 6.5390e-01
#>   9  4.8172e-01 5.7815e-01 5.0937e-01 5.2736e-01 4.3290e-01 5.6358e-01
#>   10 4.6844e-01 4.0032e-01 6.4041e-01 4.1830e-01 5.0100e-01 6.0484e-01
#>     year
#> age  1999       2000       2001       2002       2003       2004      
#>   1  9.0945e+04 1.2537e+05 8.4848e+04 2.7233e+05 9.0858e+04 2.3497e+05
#>   2  1.7948e+05 1.2688e+05 1.6986e+05 1.5811e+05 4.8988e+05 1.3787e+05
#>   3  3.0039e+05 8.1844e+04 1.0835e+05 1.1273e+05 7.1881e+04 1.6115e+05
#>   4  4.0972e+04 6.6408e+04 5.2467e+04 1.5641e+04 2.0089e+04 9.2480e+03
#>   5  3.0075e+02 1.4867e+03 3.8302e+04 2.5040e+03 3.7952e+03 2.5143e+03
#>   6  4.0576e+01 2.3794e+02 1.6046e+02 7.7415e+02 4.0583e+02 5.3217e+02
#>   7  4.6055e-01 2.5015e+02 4.4670e+01 8.6596e+01 4.6622e+03 5.7687e+01
#>   8  4.8116e-01 1.8426e+02 1.7975e+00 1.4366e+00 1.5369e+00 1.7326e+00
#>   9  5.0794e-01 6.3387e-01 5.5027e-01 3.7301e-01 4.8834e-01 5.8732e-01
#>   10 4.9442e-01 5.6436e-01 4.0171e-01 4.5858e-01 7.5218e-01 1.1666e+00
#>     year
#> age  2005       2006       2007       2008       2009       2010      
#>   1  1.6721e+05 1.4818e+05 1.8753e+05 1.5508e+05 1.6542e+05 2.0555e+05
#>   2  2.5644e+05 1.6088e+05 1.6114e+05 2.4688e+05 1.6687e+05 1.2740e+05
#>   3  5.1454e+04 1.0904e+05 8.0939e+04 5.0466e+04 8.7277e+04 6.8997e+04
#>   4  2.4396e+04 8.4389e+03 1.5849e+04 5.4949e+03 1.0242e+04 2.5115e+04
#>   5  3.5202e+03 2.3345e+03 1.2333e+03 1.8364e+03 1.9048e+03 2.8677e+03
#>   6  5.2062e+03 2.6506e+02 6.4260e+03 3.8992e+02 1.1170e+03 1.6613e+03
#>   7  2.6991e+02 7.2010e+02 2.5422e+02 7.5234e+03 2.0345e+02 2.0268e+03
#>   8  5.4120e+00 1.8027e+02 1.1820e+03 8.5812e+02 2.5561e+03 2.9822e+02
#>   9  4.1276e-01 3.1454e-01 6.1069e-01 5.2758e-01 4.7938e-01 4.2617e-01
#>   10 9.1242e-01 6.1889e-01 5.6315e-01 7.8784e-01 1.0866e+00 8.0005e-01
#>     year
#> age  2011       2012       2013       2014       2015       2016      
#>   1  1.6294e+05 9.7975e+04 1.2891e+05 2.0874e+05 1.2781e+05 1.2891e+05
#>   2  1.6196e+05 2.0172e+05 1.6594e+05 2.0420e+05 2.2353e+05 1.1000e+05
#>   3  7.2807e+04 1.2371e+05 1.4435e+05 1.2380e+05 1.3518e+05 1.5138e+05
#>   4  3.4408e+04 3.1071e+04 2.9199e+04 5.8196e+04 4.1878e+04 4.5554e+04
#>   5  1.1733e+04 9.7334e+03 5.4061e+03 1.8616e+04 1.3510e+04 1.6806e+04
#>   6  2.7743e+03 1.5592e+03 1.0302e+03 2.9574e+03 2.4681e+03 6.9240e+03
#>   7  1.6325e+02 2.4324e+02 2.2069e+02 5.3583e+02 5.9475e+02 1.9259e+03
#>   8  1.2859e+03 8.0079e+00 1.6825e+01 6.8330e+01 1.7934e+02 1.3357e+03
#>   9  3.9119e-01 3.9673e-01 4.3003e-01 4.7312e-01 4.7754e-01 2.8596e-01
#>   10 6.6941e-01 1.0649e+00 7.8479e-01 9.9108e-01 7.6460e-01 8.4887e-01
#>     year
#> age  2017      
#>   1  1.1503e+05
#>   2  1.3859e+05
#>   3  8.3966e+04
#>   4  6.6378e+04
#>   5  1.6426e+04
#>   6  4.2243e+03
#>   7  1.5008e+03
#>   8  2.6867e+02
#>   9  4.1828e-01
#>   10 1.3820e+00
#> 
#> units:  1000 
#> 
#> Slot "discards.wt":
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  1957  1958  1959  1960  1961  1962  1963  1964  1965  1966  1967  1968 
#>   1  0.044 0.047 0.051 0.045 0.044 0.042 0.048 0.032 0.038 0.038 0.036 0.060
#>   2  0.104 0.096 0.107 0.112 0.100 0.098 0.105 0.114 0.072 0.101 0.105 0.096
#>   3  0.146 0.158 0.155 0.159 0.160 0.155 0.156 0.160 0.166 0.125 0.158 0.156
#>   4  0.181 0.188 0.186 0.188 0.194 0.193 0.188 0.192 0.192 0.194 0.169 0.191
#>   5  0.206 0.200 0.197 0.204 0.204 0.213 0.205 0.204 0.212 0.205 0.220 0.192
#>   6  0.244 0.244 0.231 0.212 0.220 0.221 0.231 0.221 0.221 0.231 0.220 0.244
#>   7  0.244 0.000 0.000 0.244 0.220 0.221 0.221 0.244 0.231 0.231 0.244 0.220
#>   8  0.231 0.000 0.000 0.000 0.000 0.231 0.231 0.231 0.000 0.244 0.244 0.000
#>   9  0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#>   10 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#>     year
#> age  1969  1970  1971  1972  1973  1974  1975  1976  1977  1978  1979  1980 
#>   1  0.052 0.049 0.057 0.061 0.043 0.054 0.068 0.085 0.069 0.069 0.066 0.055
#>   2  0.146 0.114 0.110 0.147 0.131 0.106 0.136 0.153 0.160 0.143 0.158 0.149
#>   3  0.162 0.179 0.183 0.173 0.179 0.173 0.162 0.176 0.186 0.197 0.185 0.191
#>   4  0.186 0.189 0.200 0.211 0.195 0.212 0.206 0.195 0.196 0.205 0.204 0.212
#>   5  0.211 0.196 0.212 0.211 0.211 0.220 0.221 0.220 0.198 0.211 0.220 0.231
#>   6  0.212 0.000 0.000 0.244 0.244 0.231 0.244 0.000 0.220 0.213 0.231 0.000
#>   7  0.000 0.220 0.000 0.000 0.000 0.244 0.244 0.244 0.000 0.231 0.221 0.000
#>   8  0.231 0.231 0.231 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.244 0.000
#>   9  0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#>   10 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#>     year
#> age  1981  1982  1983  1984  1985  1986  1987  1988  1989  1990  1991  1992 
#>   1  0.048 0.054 0.051 0.053 0.054 0.049 0.043 0.043 0.046 0.051 0.055 0.050
#>   2  0.135 0.126 0.126 0.127 0.139 0.124 0.105 0.098 0.102 0.111 0.130 0.122
#>   3  0.179 0.182 0.180 0.172 0.177 0.181 0.166 0.153 0.163 0.157 0.161 0.167
#>   4  0.212 0.203 0.205 0.211 0.197 0.196 0.205 0.185 0.181 0.186 0.185 0.188
#>   5  0.220 0.231 0.211 0.205 0.231 0.220 0.220 0.220 0.196 0.212 0.203 0.204
#>   6  0.000 0.244 0.244 0.000 0.244 0.244 0.231 0.244 0.000 0.231 0.221 0.212
#>   7  0.000 0.244 0.000 0.244 0.000 0.244 0.000 0.000 0.000 0.000 0.231 0.231
#>   8  0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.231 0.244
#>   9  0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#>   10 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#>     year
#> age  1993  1994  1995  1996  1997  1998  1999  2000  2001  2002  2003  2004 
#>   1  0.056 0.060 0.058 0.052 0.044 0.047 0.053 0.059 0.068 0.056 0.064 0.054
#>   2  0.121 0.140 0.141 0.122 0.117 0.086 0.097 0.110 0.122 0.119 0.113 0.117
#>   3  0.171 0.175 0.186 0.179 0.178 0.170 0.143 0.151 0.167 0.170 0.174 0.164
#>   4  0.197 0.194 0.201 0.205 0.203 0.199 0.197 0.174 0.178 0.182 0.185 0.183
#>   5  0.211 0.213 0.220 0.221 0.221 0.220 0.220 0.244 0.197 0.172 0.198 0.189
#>   6  0.231 0.244 0.232 0.232 0.244 0.000 0.000 0.000 0.244 0.208 0.204 0.192
#>   7  0.244 0.244 0.232 0.000 0.000 0.244 0.000 0.203 0.000 0.003 0.221 0.196
#>   8  0.000 0.221 0.244 0.000 0.000 0.000 0.000 0.000 0.244 0.000 0.000 0.000
#>   9  0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#>   10 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#>     year
#> age  2005  2006  2007  2008  2009  2010  2011  2012  2013  2014  2015  2016 
#>   1  0.061 0.060 0.055 0.056 0.060 0.060 0.047 0.052 0.051 0.025 0.026 0.048
#>   2  0.109 0.128 0.098 0.116 0.116 0.117 0.104 0.093 0.081 0.089 0.078 0.079
#>   3  0.170 0.164 0.177 0.163 0.164 0.159 0.162 0.142 0.127 0.132 0.122 0.124
#>   4  0.175 0.193 0.178 0.186 0.199 0.199 0.171 0.188 0.151 0.162 0.149 0.150
#>   5  0.215 0.198 0.188 0.187 0.202 0.190 0.192 0.198 0.170 0.180 0.164 0.151
#>   6  0.205 0.204 0.199 0.230 0.212 0.198 0.196 0.206 0.194 0.212 0.185 0.179
#>   7  0.210 0.212 0.225 0.220 0.210 0.211 0.199 0.215 0.228 0.300 0.173 0.166
#>   8  0.176 0.220 0.200 0.191 0.220 0.234 0.211 0.215 0.346 0.370 0.218 0.192
#>   9  0.000 0.000 0.000 0.000 0.000 0.001 0.000 0.000 0.000 0.255 0.404 0.251
#>   10 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#>     year
#> age  2017 
#>   1  0.051
#>   2  0.080
#>   3  0.121
#>   4  0.139
#>   5  0.161
#>   6  0.194
#>   7  0.208
#>   8  0.206
#>   9  0.513
#>   10 0.000
#> 
#> units:  kg 
#> 
#> Slot "catch.sel":
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  1957     1958     1959     1960     1961     1962     1963     1964    
#>   1  0.319597 0.345566 0.365500 0.381717 0.385631 0.322657 0.217307 0.152594
#>   2  0.553283 0.645282 0.711574 0.728462 0.724678 0.688565 0.626367 0.571353
#>   3  0.846814 0.942805 1.000000 1.000000 1.000000 1.000000 1.000000 1.000000
#>   4  1.000000 1.000000 0.985358 0.975333 0.997282 0.974967 0.889286 0.833168
#>   5  0.829895 0.888824 0.927207 0.950057 0.982962 0.957895 0.864955 0.803482
#>   6  0.683056 0.759374 0.819713 0.864099 0.918415 0.926963 0.871700 0.822474
#>   7  0.729890 0.687432 0.658678 0.673762 0.743951 0.779122 0.727833 0.666898
#>   8  0.737425 0.677256 0.624970 0.601669 0.621474 0.628627 0.595365 0.561877
#>   9  0.643317 0.707743 0.717593 0.644573 0.561409 0.512506 0.511277 0.534749
#>   10 0.643317 0.707743 0.717593 0.644573 0.561409 0.512506 0.511277 0.534749
#>     year
#> age  1965     1966     1967     1968     1969     1970     1971     1972    
#>   1  0.144813 0.181590 0.256389 0.370866 0.478687 0.519079 0.495916 0.433839
#>   2  0.548571 0.566048 0.625079 0.739706 0.831124 0.840999 0.795301 0.718076
#>   3  1.000000 1.000000 0.988487 1.000000 1.000000 1.000000 1.000000 0.961202
#>   4  0.878555 0.977774 0.995583 0.935378 0.874962 0.899835 0.975990 1.000000
#>   5  0.847939 0.956345 1.000000 0.968135 0.914427 0.916355 0.955209 0.955395
#>   6  0.838284 0.903595 0.956088 0.995515 0.993070 0.970778 0.941919 0.882950
#>   7  0.663277 0.710328 0.765266 0.829730 0.869858 0.888381 0.877674 0.801671
#>   8  0.564655 0.597858 0.630606 0.673989 0.717216 0.774245 0.812742 0.759020
#>   9  0.558618 0.572769 0.568127 0.575453 0.603310 0.679395 0.763601 0.751839
#>   10 0.558618 0.572769 0.568127 0.575453 0.603310 0.679395 0.763601 0.751839
#>     year
#> age  1973     1974     1975     1976     1977     1978     1979     1980    
#>   1  0.400960 0.434772 0.556989 0.721753 0.805591 0.732478 0.558284 0.391028
#>   2  0.679540 0.694717 0.757884 0.839313 0.897170 0.893882 0.836701 0.745009
#>   3  0.936906 0.932586 0.936101 0.947638 0.969132 0.983763 1.000000 1.000000
#>   4  1.000000 1.000000 1.000000 1.000000 1.000000 0.981715 0.954506 0.917619
#>   5  0.950511 0.947907 0.941877 0.950096 0.987575 1.000000 0.937403 0.813353
#>   6  0.846213 0.831516 0.829200 0.853914 0.918995 0.959094 0.899456 0.751833
#>   7  0.732821 0.709530 0.734448 0.779846 0.812626 0.803713 0.759805 0.685806
#>   8  0.685916 0.662038 0.700313 0.746497 0.736930 0.676229 0.623270 0.587636
#>   9  0.696721 0.677572 0.713252 0.741681 0.695594 0.597889 0.519687 0.477253
#>   10 0.696721 0.677572 0.713252 0.741681 0.695594 0.597889 0.519687 0.477253
#>     year
#> age  1981     1982     1983     1984     1985     1986     1987     1988    
#>   1  0.320719 0.359600 0.438992 0.508227 0.525411 0.476314 0.421282 0.372567
#>   2  0.688375 0.694787 0.691201 0.702492 0.725468 0.720405 0.729258 0.720448
#>   3  1.000000 1.000000 0.935954 0.899310 0.906692 0.911941 0.963969 0.995556
#>   4  0.920094 0.983107 1.000000 1.000000 0.993830 0.958676 0.970774 0.981321
#>   5  0.747337 0.793523 0.855409 0.931447 1.000000 1.000000 1.000000 1.000000
#>   6  0.650128 0.647408 0.676164 0.764187 0.895586 0.967858 0.997996 0.999845
#>   7  0.625563 0.598767 0.574290 0.612792 0.714063 0.796867 0.858469 0.863745
#>   8  0.572297 0.571147 0.545623 0.546882 0.585184 0.630831 0.710006 0.748057
#>   9  0.482558 0.536901 0.560887 0.542507 0.510860 0.507997 0.595523 0.683599
#>   10 0.482558 0.536901 0.560887 0.542507 0.510860 0.507997 0.595523 0.683599
#>     year
#> age  1989     1990     1991     1992     1993     1994     1995     1996    
#>   1  0.337208 0.317958 0.313608 0.317364 0.308537 0.274436 0.230320 0.182277
#>   2  0.670073 0.613287 0.591325 0.608569 0.620411 0.594937 0.571485 0.566898
#>   3  0.941979 0.855419 0.811002 0.828641 0.865622 0.889307 0.933892 1.000000
#>   4  0.944005 0.882320 0.834448 0.823171 0.847719 0.905517 0.975389 0.973752
#>   5  1.000000 0.992846 0.959169 0.914715 0.899085 0.936095 0.987731 0.946386
#>   6  0.997378 1.000000 1.000000 1.000000 1.000000 1.000000 1.000000 0.948663
#>   7  0.799478 0.737914 0.758372 0.875772 0.989228 0.983437 0.918941 0.847192
#>   8  0.674489 0.587316 0.603266 0.749131 0.902479 0.890662 0.795492 0.713171
#>   9  0.644446 0.559019 0.560466 0.677526 0.795813 0.768684 0.673155 0.591838
#>   10 0.644446 0.559019 0.560466 0.677526 0.795813 0.768684 0.673155 0.591838
#>     year
#> age  1997     1998     1999     2000     2001     2002     2003     2004    
#>   1  0.143043 0.151843 0.195936 0.226516 0.257958 0.248956 0.257362 0.377474
#>   2  0.543728 0.547334 0.494571 0.422265 0.525280 0.668846 0.712655 0.777449
#>   3  1.000000 1.000000 0.856585 0.685020 0.815072 1.000000 1.000000 1.000000
#>   4  0.878155 0.910350 0.971401 0.907381 0.934026 0.850734 0.740959 0.814703
#>   5  0.819209 0.861938 1.000000 1.000000 1.000000 0.831894 0.678259 0.734727
#>   6  0.834307 0.848650 0.889383 0.840242 0.921086 0.888208 0.743459 0.709763
#>   7  0.746425 0.714361 0.646504 0.556054 0.665874 0.761966 0.674423 0.580840
#>   8  0.626964 0.597783 0.526520 0.427322 0.468024 0.493595 0.428649 0.384401
#>   9  0.520875 0.526719 0.510825 0.408282 0.347043 0.265007 0.208709 0.220789
#>   10 0.520875 0.526719 0.510825 0.408282 0.347043 0.265007 0.208709 0.220789
#>     year
#> age  2005     2006     2007     2008     2009     2010     2011     2012    
#>   1  0.508322 0.459156 0.378579 0.437099 0.635826 0.671595 0.465861 0.320466
#>   2  0.825965 0.825978 0.811439 0.828814 0.861594 0.810648 0.749683 0.692005
#>   3  1.000000 1.000000 1.000000 1.000000 1.000000 0.964482 0.996269 1.000000
#>   4  0.890681 0.810159 0.710516 0.743343 0.904035 1.000000 1.000000 0.924068
#>   5  0.790981 0.679646 0.552063 0.544056 0.648375 0.739690 0.795331 0.778818
#>   6  0.677755 0.586089 0.497958 0.468312 0.490430 0.508621 0.547891 0.557053
#>   7  0.499849 0.441687 0.412850 0.418561 0.436416 0.403059 0.353745 0.309307
#>   8  0.337920 0.283208 0.247291 0.250257 0.275155 0.260659 0.219264 0.182084
#>   9  0.221407 0.163375 0.112445 0.101091 0.117965 0.131563 0.133021 0.123273
#>   10 0.221407 0.163375 0.112445 0.101091 0.117965 0.131563 0.133021 0.123273
#>     year
#> age  2013     2014     2015     2016     2017    
#>   1  0.371582 0.574650 0.674177 0.485580 0.268807
#>   2  0.727648 0.817662 0.848662 0.775415 0.659347
#>   3  1.000000 1.000000 0.997872 1.000000 1.000000
#>   4  0.921248 0.974274 1.000000 0.966840 0.901408
#>   5  0.770034 0.777282 0.780296 0.775554 0.763304
#>   6  0.553585 0.547440 0.546097 0.556122 0.570028
#>   7  0.317959 0.360592 0.384355 0.363367 0.320835
#>   8  0.187146 0.217980 0.230463 0.203954 0.163486
#>   9  0.121437 0.123710 0.117807 0.101872 0.083357
#>   10 0.121437 0.123710 0.117807 0.101872 0.083357
#> 
#> units:   
#> 
#> Slot "price":
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  1957 1958 1959 1960 1961 1962 1963 1964 1965 1966 1967 1968 1969 1970 1971
#>   1  NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>   2  NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>   3  NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>   4  NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>   5  NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>   6  NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>   7  NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>   8  NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>   9  NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>   10 NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>     year
#> age  1972 1973 1974 1975 1976 1977 1978 1979 1980 1981 1982 1983 1984 1985 1986
#>   1  NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>   2  NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>   3  NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>   4  NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>   5  NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>   6  NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>   7  NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>   8  NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>   9  NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>   10 NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>     year
#> age  1987 1988 1989 1990 1991 1992 1993 1994 1995 1996 1997 1998 1999 2000 2001
#>   1  NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>   2  NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>   3  NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>   4  NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>   5  NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>   6  NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>   7  NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>   8  NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>   9  NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>   10 NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>     year
#> age  2002 2003 2004 2005 2006 2007 2008 2009 2010 2011 2012 2013 2014 2015 2016
#>   1  NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>   2  NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>   3  NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>   4  NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>   5  NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>   6  NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>   7  NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>   8  NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>   9  NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>   10 NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>     year
#> age  2017
#>   1  NA  
#>   2  NA  
#>   3  NA  
#>   4  NA  
#>   5  NA  
#>   6  NA  
#>   7  NA  
#>   8  NA  
#>   9  NA  
#>   10 NA  
#> 
#> units:  1000 
#> 
#> Slot "catch.q":
#> An object of class "FLPar"
#> params
#> alpha  beta 
#>     1     0 
#> units:  NA 
#> 
#> Slot "name":
#> [1] "PLE"
#> 
#> Slot "desc":
#> [1] "Plaice in IV. ICES WGNSSK 2018. FLAAP"
#> 
#> Slot "range":
#>       min       max plusgroup   minyear   maxyear 
#>         1        10        10      1957      2017 
#> 
#> 
#> Slot "capacity":
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1957 1958 1959 1960 1961 1962 1963 1964 1965 1966 1967 1968 1969 1970
#>   all NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>      year
#> quant 1971 1972 1973 1974 1975 1976 1977 1978 1979 1980 1981 1982 1983 1984
#>   all NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>      year
#> quant 1985 1986 1987 1988 1989 1990 1991 1992 1993 1994 1995 1996 1997 1998
#>   all NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>      year
#> quant 1999 2000 2001 2002 2003 2004 2005 2006 2007 2008 2009 2010 2011 2012
#>   all NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>      year
#> quant 2013 2014 2015 2016 2017
#>   all NA   NA   NA   NA   NA  
#> 
#> units:  NA 
#> 
#> Slot "effort":
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1957 1958 1959 1960 1961 1962 1963 1964 1965 1966 1967 1968 1969 1970
#>   all NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>      year
#> quant 1971 1972 1973 1974 1975 1976 1977 1978 1979 1980 1981 1982 1983 1984
#>   all NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>      year
#> quant 1985 1986 1987 1988 1989 1990 1991 1992 1993 1994 1995 1996 1997 1998
#>   all NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>      year
#> quant 1999 2000 2001 2002 2003 2004 2005 2006 2007 2008 2009 2010 2011 2012
#>   all NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>      year
#> quant 2013 2014 2015 2016 2017
#>   all NA   NA   NA   NA   NA  
#> 
#> units:  NA 
#> 
#> Slot "hperiod":
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>        year
#> quant   1957 1958 1959 1960 1961 1962 1963 1964 1965 1966 1967 1968 1969 1970
#>   start 0    0    0    0    0    0    0    0    0    0    0    0    0    0   
#>   end   1    1    1    1    1    1    1    1    1    1    1    1    1    1   
#>        year
#> quant   1971 1972 1973 1974 1975 1976 1977 1978 1979 1980 1981 1982 1983 1984
#>   start 0    0    0    0    0    0    0    0    0    0    0    0    0    0   
#>   end   1    1    1    1    1    1    1    1    1    1    1    1    1    1   
#>        year
#> quant   1985 1986 1987 1988 1989 1990 1991 1992 1993 1994 1995 1996 1997 1998
#>   start 0    0    0    0    0    0    0    0    0    0    0    0    0    0   
#>   end   1    1    1    1    1    1    1    1    1    1    1    1    1    1   
#>        year
#> quant   1999 2000 2001 2002 2003 2004 2005 2006 2007 2008 2009 2010 2011 2012
#>   start 0    0    0    0    0    0    0    0    0    0    0    0    0    0   
#>   end   1    1    1    1    1    1    1    1    1    1    1    1    1    1   
#>        year
#> quant   2013 2014 2015 2016 2017
#>   start 0    0    0    0    0   
#>   end   1    1    1    1    1   
#> 
#> units:  NA 
#> 
#> Slot "vcost":
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1957 1958 1959 1960 1961 1962 1963 1964 1965 1966 1967 1968 1969 1970
#>   all NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>      year
#> quant 1971 1972 1973 1974 1975 1976 1977 1978 1979 1980 1981 1982 1983 1984
#>   all NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>      year
#> quant 1985 1986 1987 1988 1989 1990 1991 1992 1993 1994 1995 1996 1997 1998
#>   all NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>      year
#> quant 1999 2000 2001 2002 2003 2004 2005 2006 2007 2008 2009 2010 2011 2012
#>   all NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>      year
#> quant 2013 2014 2015 2016 2017
#>   all NA   NA   NA   NA   NA  
#> 
#> units:  NA 
#> 
#> Slot "fcost":
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1957 1958 1959 1960 1961 1962 1963 1964 1965 1966 1967 1968 1969 1970
#>   all NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>      year
#> quant 1971 1972 1973 1974 1975 1976 1977 1978 1979 1980 1981 1982 1983 1984
#>   all NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>      year
#> quant 1985 1986 1987 1988 1989 1990 1991 1992 1993 1994 1995 1996 1997 1998
#>   all NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>      year
#> quant 1999 2000 2001 2002 2003 2004 2005 2006 2007 2008 2009 2010 2011 2012
#>   all NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>      year
#> quant 2013 2014 2015 2016 2017
#>   all NA   NA   NA   NA   NA  
#> 
#> units:  NA 
#> 
#> Slot "orevenue":
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1957 1958 1959 1960 1961 1962 1963 1964 1965 1966 1967 1968 1969 1970
#>   all NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>      year
#> quant 1971 1972 1973 1974 1975 1976 1977 1978 1979 1980 1981 1982 1983 1984
#>   all NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>      year
#> quant 1985 1986 1987 1988 1989 1990 1991 1992 1993 1994 1995 1996 1997 1998
#>   all NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>      year
#> quant 1999 2000 2001 2002 2003 2004 2005 2006 2007 2008 2009 2010 2011 2012
#>   all NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  
#>      year
#> quant 2013 2014 2015 2016 2017
#>   all NA   NA   NA   NA   NA  
#> 
#> units:  NA 
#> 
#> Slot "crewshare":
#> An object of class "FLQuants": EMPTY
#> model:  
#> ~fixed * crew + share * lrevenue
#> <environment: namespace:FLFishery>
#> 
#> params:  
#> An object of class "FLPar"
#> params
#> fixed  crew share 
#>     0     1     0 
#> units:  NA 
#> 
#> Slot "name":
#> [1] ""
#> 
#> Slot "desc":
#> [1] ""
#> 
#> Slot "range":
#>     min     max minyear maxyear 
#>       0       0       1       1 
#> 
#> Slot "names":
#> [1] "PLE"
#> 
#> Slot "lock":
#> [1] FALSE
#>